==> src/1_translate_and_describe_FR.Rmd <==
---
title: "Donor iron supplement perception study data description and translation - France"
author: "Amber Meulenbeld"
date: "`r Sys.time()`"
output:
  html_document:
    df_print: paged
  pdf_document:
    fig_caption: yes
    number_sections: yes
---

\tableofcontents


```{r setup, include=FALSE}

  
library(knitr)
opts_chunk$set(#results = 'asis',      # Can also be set at the chunk-level. IT IS NEEDED IN THE dfsummary chunck
               comment = NA,
               prompt  = FALSE,
               cache   = FALSE,
               echo = TRUE,
               message = FALSE,
               warning = FALSE,
               width.cutoff=60
)
library(summarytools)
st_options(plain.ascii = FALSE,        # Always use this option in Rmd documents
           style        = "rmarkdown", # Always use this option in Rmd documents
           footnote     = NA,          # Makes html-rendered results more concise
           subtitle.emphasis = FALSE)  # Improves layout with some rmardown themes
library(tidyverse)
library(readxl)
library(lubridate)
library(summarytools)

```


# Read and summarise data


```{r}
Data <- read_excel("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/A Source documents/iDPS - FR/iDPS - Raw data - France.xlsx")
data <- Data
colnames(data)
```


# Translate questions
```{r}

# Demographics
names(data) <- gsub("Age", "Age", names(data))
names(data) <- gsub("Gender", "Sex", names(data))
names(data) <- gsub("Do you have periods", "Menstruation", names(data))
names(data) <- gsub("Where do you live", "Area", names(data))
names(data) <- gsub("Scholarship", "Education", names(data))
names(data) <- gsub("Health rating", "Health", names(data))
names(data) <- gsub("Are you vegetarian/vegetalian", "Diet", names(data))
names(data) <- gsub("Number of donations \\(lifetime\\)", "Number_donation", names(data))
names(data) <- gsub("Last donation", "Last_donation", names(data))
#names(data) <- gsub("^Q92$", "Dem_Location_donation", names(data))
#names(data) <- gsub("^Q8_4_TEXT$", "Dem_Sex_other", names(data))
#names(data) <- gsub("^Q9_5_TEXT$", "Demo_Menstruation_other", names(data))

# Opinions blood service
names(data) <- gsub("EFS is a professional organisation", "OpinionBloodService1", names(data))
names(data) <- gsub("Opportunity to ask questions", "OpinionBloodService2", names(data))
names(data) <- gsub("BS treat my personal information with care", "OpinionBloodService3", names(data))
names(data) <- gsub("I am approached personally", "OpinionBloodService4", names(data))
names(data) <- gsub("Disapointed if I could not give blood anaymore", "OpinionBloodService5", names(data))
names(data) <- gsub("I donate so I can monitor my health status", "OpinionBloodService6", names(data))
names(data) <- gsub("I trust the donor BS staff", "OpinionBloodService7", names(data))
names(data) <- gsub("confident that the health of donors is a high priority", "OpinionBloodService8", names(data))

# Donation-related symptoms - severity
names(data) <- gsub("Severity of bruise", "Symptoms1_Severity", names(data))
names(data) <- gsub("Severity of fainting", "Symptoms2_Severity", names(data))
names(data) <- gsub("Severity of tiredness", "Symptoms3_Severity", names(data))
names(data) <- gsub("Severity :Painful arm", "Symptoms4_Severity", names(data))
names(data) <- gsub("Severity :Feeling unwell", "Symptoms5_Severity", names(data))
names(data) <- gsub("Severity ofdizziness", "Symptoms6_Severity", names(data))
names(data) <- gsub("Severity of blledding", "Symptoms7_Severity", names(data))
names(data) <- gsub("Severity of headache", "Symptoms8_Severity", names(data))
names(data) <- gsub("Importance : Increased energy after the donation", "Symptoms9_Severity", names(data))
names(data) <- gsub("Severity :Thirst", "Symptoms10_Severity", names(data))

# Donation-related symptoms
names(data) <- gsub("Occurrende of bruise", "Symptoms1", names(data))
names(data) <- gsub("Occurrence of fainting", "Symptoms2", names(data))
names(data) <- gsub("Occurence of tiredness", "Symptoms3", names(data))
names(data) <- gsub("Painful arm", "Symptoms4", names(data))
names(data) <- gsub("Feeling unwell", "Symptoms5", names(data))
names(data) <- gsub("Occurrence of dizziness", "Symptoms6", names(data))
names(data) <- gsub("Occurence of bleedding", "Symptoms7", names(data))
names(data) <- gsub("Occurrence of headache", "Symptoms8", names(data))
names(data) <- gsub("Increased energy after the donation", "Symptoms9", names(data))
names(data) <- gsub("Thirst", "Symptoms10", names(data))

# Knowledge donation - Confidence
names(data) <- gsub("Confidence rating of answer : A donor gives 500ml of blood", "Knowledge1_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :W are allowed to donate more often than men", "Knowledge2_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :A donor loses red bloo cells", "Knowledge3_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :A donor loses iron during a donation", "Knowledge4_Certainty", names(data))

# Knowledge donation
names(data) <- gsub("A donor gives 500ml of blood", "Knowledge1", names(data))
names(data) <- gsub("W are allowed to donate more often than men", "Knowledge2", names(data))
names(data) <- gsub("A donor loses red bloo cells", "Knowledge3", names(data))
names(data) <- gsub("A donor loses iron during a donation", "Knowledge4", names(data))

# Knowledge Hb - Confidence
names(data) <- gsub("Confidence rating of answer :Hb main function is blood clotting", "KnowledgeHb1_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :Hb transports O2", "KnowledgeHb2_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :Hb is a measure of the iron store", "KnowledgeHb3_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :Hb is measured every donation", "KnowledgeHb4_Certainty", names(data))

# Knowledge Hb
names(data) <- gsub("Hb main function is blood clotting", "KnowledgeHb1", names(data))
names(data) <- gsub("Hb transports O2", "KnowledgeHb2", names(data))
names(data) <- gsub("Hb is a measure of the iron store", "KnowledgeHb3", names(data))
names(data) <- gsub("Hb is measured every donation", "KnowledgeHb4", names(data))

# Knowledge Ferritin - Confidence
names(data) <- gsub("Confidence rating of answer :Iron is needed for the production of HB", "KnowledgeFer1_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :Iron transports O2", "KnowledgeFer2_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :Ferritin is a measure of iron stores", "KnowledgeFer3_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :Ferritin is measured every donation", "KnowledgeFer4_Certainty", names(data))

# Knowledge Ferritin
names(data) <- gsub("Iron is needed for the production of HB", "KnowledgeFer1", names(data))
names(data) <- gsub("Iron transports O2", "KnowledgeFer2", names(data))
names(data) <- gsub("Ferritin is a measure of iron stores", "KnowledgeFer3", names(data))
names(data) <- gsub("Ferritin is measured every donation", "KnowledgeFer4", names(data))

# Knowledge measurements
names(data) <- gsub("is Hb measured", "Measurement1", names(data))
names(data) <- gsub("Is transferrin measured", "Measurement2", names(data))
names(data) <- gsub("is ferritin measured", "Measurement3", names(data))
names(data) <- gsub("is hepcidin measured", "Measurement4", names(data))
names(data) <- gsub("is dopamine measured", "Measurement5", names(data))

# Consequences low Hb/Ferritin
names(data) <- gsub("What happens if a donor has a low Hb value", "ConsequenceHb", names(data))
names(data) <- gsub("What happens if a donor has a low ferritin level", "ConsequenceFer", names(data))

# Prior deferral
names(data) <- gsub("Have you ever been deferred due to low ferritin or HB", "Deferral", names(data))
names(data) <- gsub("Have you still deferred", "Deferral_Currently", names(data))
data <- data %>% select(-"Deferral_Currently") #removed because is already contained in last_deferral
names(data) <- gsub("When have you been defererred", "Last_Deferral", names(data))
names(data) <- gsub("The information about deferral was understandable", "DeferralInfo1", names(data))
names(data) <- gsub("Informative", "DeferralInfo2", names(data))
names(data) <- gsub("Memorable", "DeferralInfo3", names(data))
names(data) <- gsub("Useful", "DeferralInfo4", names(data))
names(data) <- gsub("Clear", "DeferralInfo5", names(data))

names(data) <- gsub("What information was missing", "Deferral_Info_Lacking", names(data))

# Effect deferral on donor motivation
names(data) <- gsub("The deferral put me off donating again", "Deferral_Motivation1", names(data))
names(data) <- gsub("The deferral had no effect", "Deferral_Motivation2", names(data))
names(data) <- gsub("The defrerral made me want to come back to donate again", "Deferral_Motivation3", names(data))
names(data) <- gsub("Iif you were deferred would that put you off donating again", "Deferral_Motivation_Hypothetical1", names(data))
names(data) <- gsub("Would have no effect", "Deferral_Motivation_Hypothetical2", names(data))
names(data) <- gsub("Would make you want to come back again", "Deferral_Motivation_Hypothetical3", names(data))

# Symptoms ID
names(data) <- gsub("symptoms of iron deficiency", "SymptomsID", names(data))

# Account for post-donation iron loss
names(data) <- gsub("do you do anything before you donate to help to mamge any iron loss", "AccForIronLoss", names(data))
names(data) <- gsub("How do you take into account the amount of iron you lose", "AccForIronLoss_Text", names(data))

# Deferral information sources
names(data) <- gsub("What information source have you used..", "DeferralSource", names(data))
names(data) <- gsub("Trust about the information from the EFS", "DeferralSource1_Confidence.FI", names(data))
names(data) <- gsub("From you GP", "DeferralSource2_Confidence.FI", names(data))
names(data) <- gsub("From BS website", "DeferralSource3_Confidence.FI", names(data))
names(data) <- gsub("From other internet sources", "DeferralSource4_Confidence.FI", names(data))
names(data) <- gsub("Others", "DeferralSource5_Confidence.FI", names(data))

# Attitude toward iron supplements (should have been supplements in general)
names(data) <- gsub("I think iron supplement are bad good", "IronSupplements1", names(data))
names(data) <- gsub("detrimental beneficial", "IronSupplements2", names(data))
names(data) <- gsub("to be discourgaed encouraged", "IronSupplements3", names(data))
names(data) <- gsub("worhless worhtwhile", "IronSupplements4", names(data))

# Prior supplement use
names(data) <- gsub("Did do you take dietary supplements", "SupplementsUse", names(data))
names(data) <- gsub("Multivitamins", "SupplementsUse_Type_A", names(data))
names(data) <- gsub("Specific vitamin mineral", "SupplementsUse_Type_B", names(data))
names(data) <- gsub("Other...100", "SupplementsUse_Type_C", names(data))
names(data) <- gsub("Why do you take dietary supplements", "SupplementsUse_Reason", names(data))
  names(data) <- gsub("Other reason", "SupplementsUse_Reason_Other", names(data))
  names(data) <- gsub("Namely the reason why", "SupplementsUse_Reason_Other_Text", names(data))

# Attitude towards dietary supplements  
names(data) <- gsub("open to use dietary supplements", "SupplementsOpen", names(data))
names(data) <- gsub("If yes, what circumstances", "SupplementsOpen1", names(data))
names(data) <- gsub("If no, why wouldn't you\\?", "SupplementsOpen2", names(data))

# Kmowledge iron supplements
names(data) <- gsub("Confidence rating of answer :Iron supplements are provided by the BS …", "IronSuppKnowledge1_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :Iron supplements can improve recovery", "IronSuppKnowledge2_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :Iron supplements are only available with a doctor'sprescription", "IronSuppKnowledge3_Certainty", names(data))
names(data) <- gsub("Confidence rating of answer :Iron supplements may have side effects", "IronSuppKnowledge4_Certainty", names(data))
names(data) <- gsub("Iron supplements are provided by the BS …", "IronSuppKnowledge1", names(data))
names(data) <- gsub("Iron supplements can improve recovery", "IronSuppKnowledge2", names(data))
names(data) <- gsub("Iron supplements are only available with a doctor'sprescription", "IronSuppKnowledge3", names(data))
names(data) <- gsub("Iron supplements may have side effects", "IronSuppKnowledge4", names(data))


# Side effect iron supplements
names(data) <- gsub("Do you know the posssible side effects", "IronSuppSE", names(data))
names(data) <- gsub("Have you ever taken iron supplements", "IronSuppUse", names(data))
names(data) <- gsub("Impact on your physical activity", "IronSuppEffect1", names(data))
names(data) <- gsub("On tiredness", "IronSuppEffect2", names(data))
names(data) <- gsub("On dizziness", "IronSuppEffect3", names(data))
names(data) <- gsub("On headaches", "IronSuppEffect4", names(data))
names(data) <- gsub("On GI problems", "IronSuppEffect5", names(data))
names(data) <- gsub("Other...122", "IronSuppEffect6B", names(data))
names(data) <- gsub("...123", "IronSuppEffect6A", names(data))

# Concerns iron supplements
names(data) <- gsub("Do you have any concerns about iron supplements", "Concern_IronSupp", names(data))
names(data) <- gsub("Do you have other concerns about iron supplements", "Concern_IronSupp_Other", names(data))
names(data) <- gsub("If yes, what cooncerns", "Concern_IronSupp_Other_Text", names(data))

# Perception
names(data) <- gsub("If the EFS would offer iron supplements, then", "IronSupp_BloodService", names(data))

# Reasoning
names(data) <- gsub("Your most important reason to take iron supplement", "IronSupp_Reasoning_Use", names(data))

# Preference
names(data) <- gsub("A longer interval", "IronDefPref1", names(data))
names(data) <- gsub("Dietary advice", "IronDefPref2", names(data))
names(data) <- gsub("Provision of iron supplements", "IronDefPref3", names(data))
names(data) <- gsub("Stop donations", "IronDefPref4", names(data))
names(data) <- gsub("NOe", "IronDefPref5", names(data))
# Circumstance iron supplement use
names(data) <- gsub("In what situation would you be willing to use iron supplements", "IronSupp_Circumstance", names(data))

# Most important reason for providing iron supplements
names(data) <- gsub("What do you think is most importnat whan considerng iron supplementation…", "IronSupp_Reasoning_Provision", names(data))

data <- data %>% select(-c("Suppose it turns out that your iron level is too low which of the following options would you prefer", "IronSupp_Reasoning_Provision")) #removed because question was not asked in the right manner and everyone gave the same answer?
```


# Translate answers

```{r}
categorize_education <- function(education) {
  if (education %in% c("None")) {
    return("Primary School")
  } else if (education %in% c("Bac + 2", "bac+2 en lycée", "Lycée")) {
    return("High School")
  } else if (education %in% c("Collège", "École d’infirmière L2")) {
    return("College")
  } else if (education %in% c("BTS", "BEP", "bts")) {
    return("Applied Sciences")
  } else if (education %in% c("Licence", "Bac + 3", "Bac + 5", "Doctorat", "bac +3", "Bac + 3, diplôme d’état IDE", "École ingénieurs (Bac+5)", "Université")) {
    return("University")
  } else if (education %in% c("Employée")) {
    return("Other")
  } else {
    return("Other")
  }
}

# Apply the function to the Education column
data <- data %>%
  mutate(Education = sapply(Education, categorize_education))

# Diet

translate_diet <- function(diet) {
  if (diet == "YES, vegane") {
    return("Yes, vegan (no animal products)")
  } else if (diet == "YES, végétarien (pas de viende et pas de poisson)") {
    return("Yes, vegetarian (no meat and no fish)")
  } else if (diet == "YES, pescétarien (pas de viande)") {
    return("Yes, pescatarian (no meat)")
  } else if (diet == "YES, flexitarien (au moins trois jours sans viande par semaine)") {
    return("Yes, flexitarian (at least three days without meat)")
  } else if (diet == "NO") {
    return("No")
  } else {
    return(NA)
  }
}

# Apply the function to the Diet column
data <- data %>%
  mutate(Diet = sapply(Diet, translate_diet))

# last donation
data <- data %>% mutate(Last_donation = ifelse(Last_donation == "Je ne me souviens pas", "I cannot remember", Last_donation))

#Deferral
translate_deferral <- function(deferral) {
  if (is.na(deferral)){
    return(NA)
  } else if (deferral == "Choix1") {
    return("Yes, once")
  } else if (deferral == "Choix2") {
    return("Yes, multiple times")
  } else if (deferral == "Choix3") {
    return("No")
  } else if (deferral == "Choix4") {
    return("I don't know")
  } else {
    return(NA)
  }
}

data <- data %>%
  mutate(Deferral = sapply(Deferral, translate_deferral))


#last deferral
translate_lastdeferral <- function(lastdeferral) {
  if (is.na(lastdeferral)){
    return(NA)
  } else if (lastdeferral == "I am deferred by now") {
    return("Currently")
  } else if (lastdeferral == "Within the last 6 months") {
    return("Within the last 6 months")
  } else if (lastdeferral == "6 months to a year ago.") {
    return("6 months to a year ago")
  } else if (lastdeferral == "1-2 years ago.") {
    return("1-2 years ago")
  } else if (lastdeferral == "Longer than 2 years ago.") {
    return("More than 2 years ago")
  } else if (lastdeferral == "I don't remember"){
    return("I cannot remember")
  } else {
    return(NA)
  }
}

data <- data %>%
  mutate(Last_Deferral = sapply(Last_Deferral, translate_lastdeferral))


#lack deferral info 
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    LackDeferralInfo1 = sapply(as.character(Deferral_Info_Lacking), function(x) ifelse(grepl("Informations sur ce qu'est l'hémoglobine et/ou la ferritine\\.", x), 1, 0)),
    LackDeferralInfo5 = sapply(as.character(Deferral_Info_Lacking), function(x) ifelse(grepl("Informations sur les conséquences d'un faible taux d'hémoglobine et/ou de ferritine\\.", x), 1, 0)),
    LackDeferralInfo2 = sapply(as.character(Deferral_Info_Lacking), function(x) ifelse(grepl("Informations sur la façon de rétablir un taux d'hémoglobine et/ou de ferritine normal\\.", x), 1, 0)),
    LackDeferralInfo3 = sapply(as.character(Deferral_Info_Lacking), function(x) ifelse(grepl("Nonee information ne manquait\\.", x), 1, 0)),
    
    # Capture any open text answers that do not match predefined options
    LackDeferralInfoTEXT = sapply(as.character(Deferral_Info_Lacking), function(x) {
      if (is.na(x)) {
        return(NA)
      }
      predefined_patterns <- c("Informations sur ce qu'est l'hémoglobine et/ou la ferritine\\.", 
                               "Informations sur les conséquences d'un faible taux d'hémoglobine et/ou de ferritine\\.", 
                               "Informations sur la façon de rétablir un taux d'hémoglobine et/ou de ferritine normal\\.", 
                               "Nonee information ne manquait\\.")
      # Remove predefined patterns from the string
      remaining_text <- trimws(gsub(paste(predefined_patterns, collapse = "|"), "", x))
      if (nchar(remaining_text) > 0) {
        return(remaining_text)
      } else {
        return(NA)
      }
    }),
    LackDeferralInfo5 = sapply(as.character(LackDeferralInfoTEXT), function(x) {
      if(is.na(x)){
      return(0)
    } else if (nchar(x) > 0){
      return(1)
    }
    }) 
  )

data <- data %>% select(-"Deferral_Info_Lacking")

#transform all "not applicable" text to NA
data[] <- lapply(data, function(x) ifelse(x == 'Not applicable', NA, x))

#Opinion blood service column, recode "5 4" to 5
data[names(data) %in% grep("OpinionBloodService", names(data), value = TRUE)] <- 
  lapply(data[names(data) %in% grep("OpinionBloodService", names(data), value = TRUE)], 
         function(x) ifelse(x == '5 4', '5', x))

#Certainty columns
data[grep("_Certainty", names(data))] <- lapply(data[grep("_Certainty", names(data))], as.character)

#Consequences Hb
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    ConsequenceHb1 = sapply(as.character(ConsequenceHb), function(x) ifelse(grepl("Donor is allowed to donate, but the volume of donation is reduced\\.", x), 1, 0)),
    ConsequenceHb2 = sapply(as.character(ConsequenceHb), function(x) ifelse(grepl("Donor is not allowed to donate, and is deferred for 3 monts\\.", x), 1, 0)),
    ConsequenceHb3 = sapply(as.character(ConsequenceHb), function(x) ifelse(grepl("Donor is not allowed to donate, and is deferred for 12 months\\.", x), 1, 0)),
    ConsequenceHb4 = sapply(as.character(ConsequenceHb), function(x) ifelse(grepl("Donor is not allowed to donate, and receive nutritional advice\\.", x), 1, 0)),
    ConsequenceHb5 = sapply(as.character(ConsequenceHb), function(x) ifelse(grepl("Donor is not allowed to donate, and is given iron siupplements\\.", x), 1, 0)),
    ConsequenceHb6 = sapply(as.character(ConsequenceHb), function(x) ifelse(grepl("Hb is not measured\\.", x), 1, 0)),
    ConsequenceHb7 = sapply(as.character(ConsequenceHb), function(x) ifelse(grepl("I dont know\\.", x), 1, 0))
  )

data <- data %>% select(-"ConsequenceHb")


#Consequences Fer
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    ConsequenceFer1 = sapply(as.character(ConsequenceFer), function(x) ifelse(grepl("Donor is allowed to donate, but the volume of donation is reduced\\.", x), 1, 0)),
    ConsequenceFer2 = sapply(as.character(ConsequenceFer), function(x) ifelse(grepl("Donor is not allowed to donate, and is deferred for 3 monts\\.", x), 1, 0)),
    ConsequenceFer3 = sapply(as.character(ConsequenceFer), function(x) ifelse(grepl("Donor is not allowed to donate, and is deferred for 12 months\\.", x), 1, 0)),
    ConsequenceFer4 = sapply(as.character(ConsequenceFer), function(x) ifelse(grepl("Donor is not allowed to donate, and receive nutritional advice\\.", x), 1, 0)),
    ConsequenceFer5 = sapply(as.character(ConsequenceFer), function(x) ifelse(grepl("Donor is not allowed to donate, and is given iron siupplements\\.", x), 1, 0)),
    ConsequenceFer6 = sapply(as.character(ConsequenceFer), function(x) ifelse(grepl("Ferritin is not measured\\.", x), 1, 0)),
    ConsequenceFer7 = sapply(as.character(ConsequenceFer), function(x) ifelse(grepl("I dont know\\.", x), 1, 0))
  )

data <- data %>% select(-"ConsequenceFer")

#Symptoms
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    SymptomsID1 = sapply(as.character(SymptomsID), function(x) ifelse(grepl("Restless syndrome", x), 1, 0)),
    SymptomsID2 = sapply(as.character(SymptomsID), function(x) ifelse(grepl("Dark stools", x), 1, 0)),
    SymptomsID3 = sapply(as.character(SymptomsID), function(x) ifelse(grepl("Tiredness", x), 1, 0)),
    SymptomsID4 = sapply(as.character(SymptomsID), function(x) ifelse(grepl("Stomach discomfort_pain", x), 1, 0)),
    SymptomsID5 = sapply(as.character(SymptomsID), function(x) ifelse(grepl("Pica", x), 1, 0)),
    SymptomsID6 = sapply(as.character(SymptomsID), function(x) ifelse(grepl("Headache", x), 1, 0)),
    SymptomsID7 = sapply(as.character(SymptomsID), function(x) ifelse(grepl("None of these", x), 1, 0)),
    SymptomsID8 = sapply(as.character(SymptomsID), function(x) ifelse(grepl("I dont know", x), 1, 0))
  )

data <- data %>% select(-"SymptomsID")

#Deferral source
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    DeferralSource1 = sapply(as.character(DeferralSource), function(x) ifelse(grepl("EFS", x), 1, 0)),
    DeferralSource3 = sapply(as.character(DeferralSource), function(x) ifelse(grepl("GP", x), 1, 0)),
    DeferralSource4 = sapply(as.character(DeferralSource), function(x) ifelse(grepl("EFS Website ou EFS flyer", x), 1, 0)),
    DeferralSource5 = sapply(as.character(DeferralSource), function(x) ifelse(grepl("Other Website", x), 1, 0)),
    DeferralSource7 = sapply(as.character(DeferralSource), function(x) ifelse(grepl("None", x, ignore.case = T), 1, 0)),
    
        DeferralSource6B = sapply(as.character(DeferralSource), function(x) {
      if (is.na(x)) {
        return(NA)
      }
      predefined_patterns <- c("EFS", 
                               "GP", 
                               "EFS Website ou EFS flyer", 
                               "Other Website",
                               "None",
                               "none")
      # Remove predefined patterns from the string
      remaining_text <- trimws(gsub(paste(predefined_patterns, collapse = "|"), "", x))
      if (nchar(remaining_text) > 0) {
        return(remaining_text)
      } else {
        return(NA)
      }
    }),
    DeferralSource6A = sapply(as.character(DeferralSource6B), function(x) {
      if(is.na(x)){
      return(0)
    } else if (nchar(x) > 0){
      return(1)
    }
    })
  )

data <- data %>% select(-"DeferralSource")

#SupplementsReason
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    SupplementsReason1 = sapply(as.character(SupplementsUse_Reason), function(x) ifelse(grepl("It has been prescribed by a health professional", x), 1, 0)),
    SupplementsReason2 = sapply(as.character(SupplementsUse_Reason), function(x) ifelse(grepl("Because of specific health condition", x), 1, 0)),
    SupplementsReason3 = sapply(as.character(SupplementsUse_Reason), function(x) ifelse(grepl("To have a better health", x), 1, 0)),
    SupplementsReason4 = sapply(as.character(SupplementsUse_Reason), function(x) ifelse(grepl("because of my dietary habits", x), 1, 0)),
    SupplementsReason5 = sapply(as.character(SupplementsUse_Reason), function(x) ifelse(grepl("because I donate blood", x), 1, 0)),
    SupplementsReason6A = sapply(as.character(SupplementsUse_Reason_Other), function(x) ifelse(grepl("YES", x), 1, 0)),
    
        SupplementsReason6B = SupplementsUse_Reason_Other_Text
  )

data <- data %>% select(-c("SupplementsUse_Reason", "SupplementsUse_Reason_Other", "SupplementsUse_Reason_Other_Text"))

#account for iron loss
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    AccForIronLoss1 = sapply(as.character(AccForIronLoss_Text), function(x) ifelse(grepl("diet", x), 1, 0)),
    AccForIronLoss2 = sapply(as.character(AccForIronLoss_Text), function(x) ifelse(grepl("pultivitamins", x), 1, 0)),
    AccForIronLoss3 = sapply(as.character(AccForIronLoss_Text), function(x) ifelse(grepl("supplements", x), 1, 0)),
    AccForIronLoss4B = sapply(as.character(AccForIronLoss_Text), function(x) {
      if (is.na(x)) {
        return(NA)
      }
      predefined_patterns <- c("diet", 
                               "pultivitamins", 
                               "supplements", "Nothing")
      # Remove predefined patterns from the string
      if (!grepl(paste(predefined_patterns, collapse='|'), x)) {
        return(x)
      } else {
        return(NA)
      }
    }),
    AccForIronLoss4A = sapply(as.character(AccForIronLoss4B), function(x) {
      if(is.na(x)){
      return(0)
    } else {
      return(1)
    }
    })
  )

data <- data %>% select(-"AccForIronLoss_Text")


#Iron supplements side effects 
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    IronSuppSE1 = sapply(as.character(IronSuppSE), function(x) ifelse(grepl("Restless syndrome", x), 1, 0)),
    IronSuppSE2 = sapply(as.character(IronSuppSE), function(x) ifelse(grepl("Dark stools", x), 1, 0)),
    IronSuppSE3 = sapply(as.character(IronSuppSE), function(x) ifelse(grepl("Tiredness", x), 1, 0)),
    IronSuppSE4 = sapply(as.character(IronSuppSE), function(x) ifelse(grepl("Stomach discomfort_pain", x), 1, 0)),
    IronSuppSE5 = sapply(as.character(IronSuppSE), function(x) ifelse(grepl("Pica", x), 1, 0)),
    IronSuppSE6 = sapply(as.character(IronSuppSE), function(x) ifelse(grepl("Headache", x), 1, 0)),
    IronSuppSE7 = sapply(as.character(IronSuppSE), function(x) ifelse(grepl("Tous les effets secondaires qui précèdent", x), 1, 0)),
    IronSuppSE8 = sapply(as.character(IronSuppSE), function(x) ifelse(grepl("None of these", x), 1, 0)),
    IronSuppSE9 = sapply(as.character(IronSuppSE), function(x) ifelse(grepl("I dont know", x), 1, 0))
  )

data <- data %>% select(-"IronSuppSE")

#Iron Supplement Concerns
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    ConcernIronSupp1 = sapply(as.character(Concern_IronSupp), function(x) ifelse(grepl("I dont know if iron supplements can be harmfull", x), 1, 0)),
    ConcernIronSupp2 = sapply(as.character(Concern_IronSupp), function(x) ifelse(grepl("I will not take iron supplements unless I have symptoms", x), 1, 0)),
    ConcernIronSupp3 = sapply(as.character(Concern_IronSupp), function(x) ifelse(grepl("I think that iron supplements are harmfull", x), 1, 0)),
    ConcernIronSupp4 = sapply(as.character(Concern_IronSupp), function(x) ifelse(grepl("I prefer to eat more iron rather than taking iron supplements", x), 1, 0)),
    ConcernIronSupp5A = sapply(as.character(Concern_IronSupp_Other), function(x) ifelse(grepl("YES", x), 1, 0)),
    
        ConcernIronSupp5B = Concern_IronSupp_Other_Text
  )

data <- data %>% select(-c("Concern_IronSupp", "Concern_IronSupp_Other", "Concern_IronSupp_Other_Text"))

#Iron supplementation reasoning use

data <- data %>%
  mutate(
    IronSuppSanqReason = case_when(
      IronSupp_Reasoning_Use == "To be healthier ma santé." ~ 1,
      IronSupp_Reasoning_Use == "Etre capable de donner sans augmenter les délais entre deux dons." ~ 2,
      IronSupp_Reasoning_Use == "Je ne les prendrais Never." ~ 4,
      TRUE ~ 3 # Assign 3 for "Other"
    ),
    IronSuppSanqReasonB = case_when(
      IronSuppSanqReason == 3 ~ IronSupp_Reasoning_Use,  # Only retain text if it's classified as "Other"
      TRUE ~ NA_character_  # Assign NA otherwise
    )
  )

data <- data %>% select(-c("IronSupp_Reasoning_Use"))

#IronSupp_Condition
data <- data %>%
  mutate(
    IronSupp_Condition = case_when(
      IronSupp_Circumstance == "I would not take them." ~ "I would not be willing to use them",
      IronSupp_Circumstance == "If I had symptoms taht can be tretaed by iron supplements" ~ "If I would have health complaints which could be resolved by this",
      IronSupp_Circumstance == "If it allows me to continue donating régulièrement." ~ "If this allows me to keep donating regularly",
      IronSupp_Circumstance == "If it is scinetifically approved that it increases my iron store" ~ "If there is scientific evidence that it improves my iron stores",
      IronSupp_Circumstance == "If the EFS doctor asks me to take iron" ~ "If the donor physician advises me to do so",
      IronSupp_Circumstance == "To avoid iron deficiency" ~ "To prevent potential health problems in the future"
      
    )
  )

data <- data %>% select(-c("IronSupp_Circumstance"))
```


## Mapping functions 

```{r}
#

#col 11 and 12 

f1 <- function (x){
  x <- fct_recode(x,  
                  "Disagree" = "2",
                  "Neutral"="3",
                  "Agree"="4",
                  "Totally disagree" =  "1",
                  "Totally agree" = "5"
  )
  x <- factor(
    x,
    levels = c("Totally disagree","Disagree","Neutral","Agree","Totally agree"),
    ordered = TRUE
  )
  x
}

f2 <- function (x){
  x <- fct_recode(x,  
                  "Never" = "Never",
                  "Sometimes" = "Sometimes",
                  "Often" = "Often",
                  "Always" = "Always"
  )
  x <- factor(
    x,
    levels = c("Never","Sometimes","Often","Always"),
    ordered = TRUE
  )
  x  
}

f3 <- function (x){
  x <- fct_recode(x,  
                  "Yes" = "YES",
                  "No" = "NO",
                  "I don't know" = "NA"
                  )
  x <- factor(
    x,
    levels=c("Yes","No","I don't know"),
    ordered = TRUE
  )
  x
}



f4 <- function (x){
  x <- fct_recode(x,
                  "FALSE" = "False",
                  "TRUE" = "True",
                  "I don't know" = "I dont know"
                  )
  x <- factor(
    x,
    levels=c("FALSE","TRUE"),
  )
  x
}

f5 <- function (x){
  x <- fct_recode(x, 
    "50/50" = "0.5",
    "100%" = "1"
  )
  x <- factor(
    x,
    levels = c("50/50","100%"),
    ordered = TRUE
  )
  x
}




f6 <- function (x){
  x <- fct_recode(x,
                  "No" = "NO",
                  "Yes" = "YES"
                  )
  x <- factor(
    x,
    levels=c("Yes","No"),
  )
  x
}


f7 <- function (x){
  x <- as.character(x)
  x <- fct_recode(x,  
                  "FALSE" = "0",
                  "TRUE" = "1"
                  )
  x <- factor(
    x,
    levels=c("FALSE","TRUE"),
  )
  x
}

f8 <- function (x){
  x <- fct_recode(x,  
                  "Increased" = "Increased",
                  "No effect" = "No effect",
                  "Decreased" = "Decreased"
  )
  x <- factor(
    x,
    levels=c("Increased","No effect","Decreased"),
  )
  x
}

f9 <- function (x){
  x <- fct_recode(x,
                  "Agree" = "something_in_some_language",
                  "Disagree" = "something_in_some_language"
                  
                  )
  x <- factor(
    x,
    levels=c("Agree","Disagree"),
  )
  x
}

f10 <- function (x){
  x <- factor(
    x,
    levels=c("1","2","3","4","5","6"),
    ordered = TRUE
  )
  x
}

f11 <- function (x){
  x <- factor(
    x,
    levels=c("1","2","3","4","5"),
    ordered = TRUE
  )
  x
}



f13 <- function (x){
  x <- as.character(x)
  x <- fct_recode(x,  
                  "FALSE" = "False",
                  "TRUE" = "True",
                  "I don't know" = "I dont know"
                  )
  x <- factor(
    x,
    levels=c("FALSE","TRUE", "I don't know"),
  )
  x
}

```


## Main processing

```{r}

# Columns containing just yes and no (and NA)
yesandnocols <- data %>%
  summarise(across(everything(), ~ all(.x %in% c("YES", "NO", NA), na.rm = TRUE)))

#Columns containing always, often, sometimes and never
aosncols <- data %>%
  summarise(across(everything(), ~ all(.x %in% c("Always", "Often", "Sometimes", "Never", NA), na.rm = TRUE)))

#Columns containing just 0 or 0 & 1 
zero1cols <- data %>%  
  summarise(across(everything(), ~ {
    unique_vals <- sort(unique(.x))
    all(unique_vals == 0) || all(unique_vals == c(0, 1))
  }))

#Columns containing just 1-5
# cols12345 <- data %>% summarise(across(everything(), ~ str_c(names(table(.x)),collapse="") == "12345"))
# cols2345 <- data %>% summarise(across(everything(), ~ str_c(names(table(.x)),collapse="") == "2345"))
# cols1345 <- data %>% summarise(across(everything(), ~ str_c(names(table(.x)),collapse="") == "1345"))
# cols1245 <- data %>% summarise(across(everything(), ~ str_c(names(table(.x)),collapse="") == "1245"))
# cols1235 <- data %>% summarise(across(everything(), ~ str_c(names(table(.x)),collapse="") == "1235"))
# cols1234 <- data %>% summarise(across(everything(), ~ str_c(names(table(.x)),collapse="") == "1234"))
# 
# cols12345 <- bind_rows(cols12345,
#                        cols2345,
#                        cols1345,
#                        cols1245,
#                        cols1235,
#                        cols1234) %>%  
#   summarise(across(everything(), ~ any(.x))) & !str_detect(colnames(data) ,"^49\\.\\d")

cols12345 <- data %>%
  summarise(across(everything(), ~ {
    unique_vals <- unique(na.omit(.x))
    all(unique_vals %in% 1:5) && !all(1:6 %in% unique_vals)
  }))

cols12345 <- c("Symptoms1_Severity",
"Symptoms2_Severity",
"Symptoms3_Severity",
"Symptoms4_Severity",
"Symptoms5_Severity",
"Symptoms6_Severity",
"Symptoms7_Severity",
"Symptoms8_Severity",
"Symptoms9_Severity",
"Symptoms10_Severity",
"DeferralInfo1",
"DeferralInfo2",
"DeferralInfo3",
"DeferralInfo4",
"DeferralInfo5",
"Deferral_Motivation1",
"Deferral_Motivation2",
"Deferral_Motivation3",
"Deferral_Motivation_Hypothetical1",
"Deferral_Motivation_Hypothetical2",
"Deferral_Motivation_Hypothetical3"
)

#49 has scale from 1-6. But if there are no 6s then it is detected here and ruins things later


tmp <- data %>% 
 # mutate(`something_in_some_language` = as.Date(`something_in_some_language`)) %>% 
 # mutate( `1.0: something_in_some_language` = as.numeric(`1.0: something_in_some_language`)  ) %>%
  #here a set of questions is detected with some regular expression and then one of the
  #processing functions above is applied to process those questions
  mutate_at(vars(which(str_detect(colnames(data) ,"OpinionBloodService"))), ~ f1(.) ) %>% 
  mutate_at(vars(which(as.logical(aosncols))), ~ f2(.)) %>% #symptoms columns
  #mutate_at(vars(which(str_detect(colnames(data), "Measurement"))), ~ f3(.))%>% #is already contained in yesandnocols
  mutate_at(vars(which(colnames(data)  %in% c("IronSuppKnowledge1","IronSuppKnowledge2","IronSuppKnowledge3","IronSuppKnowledge4"))), ~ f4(.) ) %>%
  mutate_at(vars(which(colnames(data) %in% c("Knowledge1","Knowledge2","Knowledge3","Knowledge4","KnowledgeHb1","KnowledgeHb2","KnowledgeHb3","KnowledgeHb4","KnowledgeFer1","KnowledgeFer2","KnowledgeFer3","KnowledgeFer4"))), ~ f13(.)) %>%
   
  mutate_at(vars(which(str_detect(colnames(data) ,"_Certainty"))), ~ f5(.) ) %>%
  #here instead of a regular expression, above a vector is made which tells which columns to process
  mutate_at(vars(which(as.logical(yesandnocols))), ~ f6(.)) %>% 
  mutate_at(vars(which(as.logical(zero1cols))), ~ f7(.) ) %>%
  mutate_at(vars(which(str_detect(colnames(data), "\\bIronSuppEffect2\\b|\\bIronSuppEffect3\\b|\\bIronSuppEffect4\\b|\\bIronSuppEffect5\\b|\\bIronSuppEffect6A\\b"))), ~ f8(.))  %>%
  #mutate_at(vars(which(colnames(data) %in% c("Symptoms1_Severity"))), ~ f11(.))
  mutate_at(vars(match(cols12345, colnames(data))), ~ f11(.)) %>%
  #mutate_at(vars(which(str_detect(colnames(data), "\\bSymptoms1_Severity\\b|\\bSymptoms2_Severity\\b|\\bSymptoms3_Severity\\b|\\bSymptoms4_Severity\\b|\\bSymptoms5_Severity\\b|\\bSymptoms6_Severity\\b|\\bSymptoms7_Severity\\b|\\bSymptoms8_Severity\\b|\\bSymptoms9_Severity\\b|\\bSymptoms10_Severity\\b|\\bDeferralInfo1\\b|\\bDeferralInfo2\\b|\\bDeferralInfo3\\b|\\bDeferralInfo4\\b|\\bDeferralInfo5\\b|\\bDeferral_Motivation1\\b|\\bDeferral_Motivation2\\b|\\bDeferral_Motivation3\\b|\\bDeferral_Motivation_Hypothetical1\\b|\\bDeferral_Motivation_Hypothetical2\\b|\\bDeferral_Motivation_Hypothetical3\\b|\\bIronSupplements1\\b|\\bIronSupplements2\\b|\\bIronSupplements3\\b|\\bIronSupplements4\\b|\\bIronSuppSanqReason\\b|\\bIronDefPref1\\b|\\bIronDefPref2\\b|\\bIronDefPref3\\b|\\bIronDefPref4\\b|\\bIronSuppSanqReason\\b|\\bDeferral_Motivation1\\b|\\bDeferral_Motivation2\\b|\\bDeferral_Motivation3\\b|\\bDeferral_Motivation_Hypothetical1\\b|\\bDeferral_Motivation_Hypothetical2\\b|\\bDeferral_Motivation_Hypothetical3\\b|"))), ~ f11(.))#%>%
  mutate(Deferral
=fct_recode(Deferral,
            "No" = "No",
            "I don’t know"= "I don't know",
            "Yes, once "= "Yes, once ",
            "Yes, multiple times "= "Yes, multiple times "
            
)%>% fct_relevel(.,
"Yes, once",
"Yes, multiple times",
"No", 
"I don't know"
      ) %>% as.ordered(.)
,
Age = fct_recode(Age,
                          "18-20"="18-20 ans",
                          "21-30"="21-30 ans",
                          "31-40"="31-40 ans",
                          "41-50"="41-50 ans",
                          "51-60"="51-60 ans",
                          "61-70"="61-70 ans"),
         Sex =  fct_recode(Sex,  
                  "Male" = "M",
                  "Female" = "F"
                  ),
    
    Area = fct_recode(Area,
                       "In a city of over 100,000 people"= "In a city of over 100,000 people",
                    "In a town of less than 100,000 people, or in the countryside"="In a city of less than 100 000 people or in the country"
                                                 
    ),
      Education =fct_recode(Education,
"None"="None" ,
"Primary school"="Primary school" , 
"High school"="High school" ,
"College"="College" ,
 "Applied sciences"="Applied sciences" ,
"University"="University" ,
"Other"="Other" 
      ) %>% fct_relevel(.,
"Other",
"None",
"Primary school", 
"High school",
"College",
"Applied sciences",
"University"
      ) %>% as.ordered(.)
,
    Health = fct_recode(    Health,
"Excellent" =  "Excellente",
"Very good" =  "Very Good",
"Good" =   "Good",
"Fair" =  "Fair",
"Poor" =  "Bad"
) %>% fct_relevel(.,
"Poor",
"Fair",
"Good",
"Very good",
"Excellent",  
) %>% as.ordered(.)
,
        Diet =
  #str_replace_all(data$Diet,"\\(|\\)","_") %>% 
    fct_recode(Diet,
        "Yes, vegan (no animal products)"="Yes, vegan (no animal products)" ,
"Yes, vegetarian (no meat and no fish)"="Yes, vegetarian (no meat and no fish)"   ,
"Yes, pescatarian (no meat)"="Yes, pescatarian (no meat)"   ,
"Yes, flexitarian (at least three days without meat)" ="Yes, flexitarian (at least three days without meat)"  ,
"No"="No" 
    ),     
    Number_donation = fct_recode(Number_donation,
  "1 - 3 times"="1-3 times" ,
 "4 - 10 times"="4-10 times" ,
"More than 10 times"="More than 10 times" ,
 "None"="None" ,
"I don't know"="I don't know"  ,
"More than 30 times"="More than 30 times" 
    ) %>% fct_relevel(.,
    "I don't know",
    "None",
    "1 - 3 times",
"4 - 10 times",
"More than 10 times",
"More than 30 times"
) %>% as.ordered(.)
,
Last_donation = fct_recode(Last_donation,
"Within the last 6 months"="Within the last 6 months"  ,
"6 months to a year ago"="6 months to a year ago"  ,
"1-2 years ago"="1-2 years ago" ,
"Longer than 2 years ago"="Longer than 2 years ago"  ,
"I cannot remember"="I cannot remember"  ,
    ) %>% fct_relevel(.,
"Within the last 6 months",
"6 months to a year ago",
"1-2 years ago",
"Longer than 2 years ago",
"I cannot remember"
                      ) %>% as.ordered(.)

,
    Last_Deferral = fct_recode(Last_Deferral,
"Currently" = "Currently",
"Within the last 6 months" =  "Within the last 6 months",
"6 months to a year ago" =  "6 months to a year ago",
"1-2 years ago" =  "1-2 years ago.",
"More than 2 years ago" = "More than 2 years ago",
"I cannot remember" = "I cannot remember",
    ) %>% fct_relevel(.,
                      "Currently",
"Within the last 6 months",
"6 months to a year ago",
"1-2 years ago",
"More than 2 years ago",
"I cannot remember"
                      ) %>% as.ordered(.)

,
SupplementsUse =
  fct_recode(SupplementsUse,
             "Never"="Never",
             "Not anymore"="Not anymore",
             "Sometimes"="Sometimes",
             "Every day"="Every day"
  )%>% fct_relevel(.,
                      "Never",
"Sometimes",
"Every day",
"Not anymore"
                      ) %>% as.ordered(.)
,
IronSuppUse =
  fct_recode(IronSuppUse,
             "No, I have never taken iron (containing) supplements"= "No I never had taken iron supplements",
             "Yes, I am currently taking iron supplements ordered to me by a docto"="Yes I am currently treated (medical prescription)",
             "Yes, I am currently taking iron supplements from a drugstore"="Yes I am taking currently iron supplements bought in pharmacy",
             "Yes, I take multivitamins that contain iron" = "Yes I am taking currently multivitamins containing iron",
             "Yes, I take multivitamins that contain iron"= "I take multivitamins but I dont know if they contain iron",
             "I have taken iron supplements/multivitamins before, but not anymore"= "I had taken iron supplements but I not currently" 
             
  ),
IronSupp_BloodService = fct_recode(as.factor(IronSupp_BloodService),
    "I think it is important that my iron levels are monitored more often" = "I feel it is important to ontrol more often my iron stores",
    "I think too much is being asked of donors" = "I feel that too much is asked to donors",
    "I think it is important that a doctor is involved" = "I think it is important to have a medical advice",
    "I would gladly make use of this" = "I will certainly appreciate to take the treatment",
    "I would have a positive attitude towards this" = "I will find that it is a good thing",
    "I would find this intrusive" = "I will find this too intrusive",
    "I would like to receive more information about this first" = "I will prefer to have more information before",
    "I would prefer this over postponing a next donation" = "I would prefer to take the treatment rather than postpone my future donations"
  )
)

  
```


# to check

```{r}

# for (i in 97:length(colnames(data))){
#   print(colnames(data)[i])
#   print(table(data[,i]))
#   print(table(tmp[,i]))
# }

```

# Deal with open text answers

## 3: might do something with the iron supplement types here later?

```{r}
#tmp %>% count(`3: something_in_some_language`) %>% arrange(desc(n)) %>% print( n = 1e3)
```




# remove the open text answers for making the summary
```{r}
tmp <- tmp %>%
  select(-c("SupplementsUse_Type_A","SupplementsUse_Type_B","SupplementsUse_Type_C","AccForIronLoss4B","LackDeferralInfoTEXT","DeferralSource6B","SupplementsReason6B","IronSuppSanqReasonB","ConcernIronSupp5B","IronSuppEffect6B", "SupplementsOpen1", "SupplementsOpen2"))
```

# Summarise 
```{r}
#Export summary count data
DataExport <- apply(tmp,MARGIN = 2,FUN = table)
saveRDS(DataExport,file="L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/2. Datasets/individual/1. iDPS - Summary data - FR.RDS")

```

```{r}
# dfSummary(data,
#           plain.ascii = FALSE, 
#           style = "grid", 
#           graph.col = TRUE,
#           graph.magnif = 0.75, 
#           valid.col = FALSE, 
#           tmp.img.dir = "/tmp")


datafr <- saveRDS(tmp, file = "L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/A Source documents/iDPS - FR/datafr.RDS")

```
