==> src/1_translate_and_describe_SWE.Rmd <==
---
title: "Donor iron supplement perception study data description and translation - Finland"
author: "Mikko Arvas"
date: "`r Sys.time()`"
output:
  html_document:
    df_print: paged
  pdf_document: 
    fig_caption: yes
    number_sections: yes
---

\tableofcontents


```{r setup, include=FALSE}

  
library(knitr)
opts_chunk$set(#results = 'asis',      # Can also be set at the chunk-level. IT IS NEEDED IN THE dfsummary chunck
               comment = NA,
               prompt  = FALSE,
               cache   = FALSE,
               echo = TRUE,
               message = FALSE,
               warning = FALSE,
               width.cutoff=60
)
library(summarytools)
st_options(plain.ascii = FALSE,        # Always use this option in Rmd documents
           style        = "rmarkdown", # Always use this option in Rmd documents
           footnote     = NA,          # Makes html-rendered results more concise
           subtitle.emphasis = FALSE)  # Improves layout with some rmardown themes
library(tidyverse)
library(readxl)
library(lubridate)
library(summarytools)
library(haven)


```


# Read and summarise data


```{r}
# Open SPSS dataset with survey data
Data <- read_xlsx("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/A Source documents/iDPS - SWE/Blood donation survey(1-183)_translation.xlsx")
data <- Data
colnames(data)
```


# Translate questions
```{r}
# Delete unused variables
data <- data[ , !(names(data) %in% c("Start time", "Completion time", "E-mail", "Name", "Last changed time"))]


names(data) <- gsub("^How old are you\\?$", "Age", names(data))
names(data) <- gsub("^Sex\\?$", "Sex", names(data))
names(data) <- gsub("^Do you have regular periods\\?$", "Menstruation", names(data))
names(data) <- gsub("^In which municipality do you live\\?$", "Area", names(data))
names(data) <- gsub("^What is your highest completed level of education\\?$", "Education", names(data))
names(data) <- gsub("^How do you feel about your health\\?$", "Health", names(data))
names(data) <- gsub("^Do you eat a vegetarian or vegan diet\\?$", "Diet", names(data))
names(data) <- gsub("^On how many occasions have you donated blood in total\\?$", "Number_donation", names(data)) 
names(data) <- gsub("^Other than today, when was your last blood donation\\?$", "Last_donation", names(data))

names(data) <- gsub("^I experience the blood center as a professional organization$", "OpinionBloodService1", names(data))
names(data) <- gsub("^When I give blood there is sufficient opportunity to ask questions$", "OpinionBloodService2", names(data))
names(data) <- gsub("^I feel that the blood center handles my personal data in a judicious manner$", "OpinionBloodService3", names(data))
names(data) <- gsub("^Staff receive me when I arrive at the blood centre$", "OpinionBloodService4", names(data))
names(data) <- gsub("^I would be disappointed if I didn\\'t get to continue giving blood$", "OpinionBloodService5", names(data))
names(data) <- gsub("^I leave blood to check my health$", "OpinionBloodService6", names(data))
names(data) <- gsub("^I trust the competence of the blood donation staff$", "OpinionBloodService7", names(data))
names(data) <- gsub("^feel that the health of blood donors is important to the Blood Centre$", "OpinionBloodService8", names(data))

#names(data) <- gsub("^Bruising at the puncture site$", "Symptoms1", names(data)) #this question is missing
names(data) <- gsub("^Fainting$", "Symptoms2", names(data))
names(data) <- gsub("^Fatigue$", "Symptoms3", names(data))
names(data) <- gsub("^Pain in the arm$", "Symptoms4", names(data))
names(data) <- gsub("^Prolonged deterioration of well\\-being \\(days\\)$", "Symptoms5", names(data))
names(data) <- gsub("^Dizziness$", "Symptoms6", names(data))
names(data) <- gsub("^Bleeding from the injection site$", "Symptoms7", names(data))
names(data) <- gsub("^Headache$", "Symptoms8", names(data))
names(data) <- gsub("^Increased energy level after donating blood$", "Symptoms9", names(data))
names(data) <- gsub("^Thirst$", "Symptoms10", names(data))


names(data) <- gsub("^Bruising at the puncture site$", "Symptoms1_Severity", names(data))
names(data) <- gsub("^Fainting2$", "Symptoms2_Severity", names(data))
names(data) <- gsub("^Fatigue2$", "Symptoms3_Severity", names(data))
#names(data) <- gsub("^Pain in the arm2$", "Symptoms4_Severity", names(data)) #this question is missing
names(data) <- gsub("^Prolonged deterioration of well-being \\(days\\)2$", "Symptoms5_Severity", names(data))
names(data) <- gsub("^Dizziness2$", "Symptoms6_Severity", names(data))
names(data) <- gsub("^Bleeding from the injection site2$", "Symptoms7_Severity", names(data))
names(data) <- gsub("^Headache2$", "Symptoms8_Severity", names(data))
names(data) <- gsub("^Increased energy level after donating blood2$", "Symptoms9_Severity", names(data))
names(data) <- gsub("^Thirst2$", "Symptoms10_Severity", names(data))

names(data) <- gsub("^During blood donation\\, 450 ml of blood is lost$", "Knowledge1", names(data))
names(data) <- gsub("^Women are allowed to donate blood more often than men$", "Knowledge2", names(data))
names(data) <- gsub("^During blood donation\\, blood donors lose red blood cells$", "Knowledge3", names(data))
names(data) <- gsub("^Blood donors lose iron in connection with blood donation$", "Knowledge4", names(data))

names(data) <- gsub("^During blood donation\\, 450 ml of blood is lost2$", "Knowledge1_Certainty", names(data))
names(data) <- gsub("^Women are allowed to donate blood more often than men2$", "Knowledge2_Certainty", names(data))
names(data) <- gsub("^During blood donation\\, blood donors lose red blood cells2$", "Knowledge3_Certainty", names(data))
names(data) <- gsub("^Blood donors lose iron in connection with blood donation2$", "Knowledge4_Certainty", names(data))

names(data) <- gsub("^Hemoglobin\\'s main function in the body is coagulation$", "KnowledgeHb1", names(data))
names(data) <- gsub("^Hemoglobin transports oxygen in the body$", "KnowledgeHb2", names(data))
names(data) <- gsub("^Hemoglobin is a measure of the body\\'s iron stores$", "KnowledgeHb3", names(data))
names(data) <- gsub("^Hemoglobin is measured at each blood donation$", "KnowledgeHb4", names(data))

names(data) <- gsub("^Hemoglobin\\'s main function in the body is coagulation2$", "KnowledgeHb1_Certainty", names(data))
names(data) <- gsub("^Hemoglobin transports oxygen in the body2$", "KnowledgeHb2_Certainty", names(data))
names(data) <- gsub("^Hemoglobin is a measure of the body\\'s iron stores2$", "KnowledgeHb3_Certainty", names(data))
names(data) <- gsub("^Hemoglobin is measured at each blood donation2$", "KnowledgeHb4_Certainty", names(data))

names(data) <- gsub("^Iron is needed for the production of the body\\'s hemoglobin$", "KnowledgeFer1", names(data))
names(data) <- gsub("^Iron transports oxygen in the body$", "KnowledgeFer2", names(data))
names(data) <- gsub("^Ferritin is a measure of the body\\'s iron stores$", "KnowledgeFer3", names(data))
names(data) <- gsub("^Ferritin is measured at each blood donation$", "KnowledgeFer4", names(data))

names(data) <- gsub("^Iron is needed for the production of the body\\'s hemoglobin2$", "KnowledgeFer1_Certainty", names(data))
names(data) <- gsub("^Iron transports oxygen in the body2$", "KnowledgeFer2_Certainty", names(data))
names(data) <- gsub("^Ferritin is a measure of the body's iron stores2$", "KnowledgeFer3_Certainty", names(data))
names(data) <- gsub("^Ferritin is measured at each blood donation2$", "KnowledgeFer4_Certainty", names(data))

names(data) <- gsub("^Hemoglobin \\(Hb\\)$", "Measurement1", names(data))
names(data) <- gsub("^Transferrin$", "Measurement2", names(data))
names(data) <- gsub("^Ferritin$", "Measurement3", names(data))
names(data) <- gsub("^Hepcidin$", "Measurement4", names(data))
names(data) <- gsub("^Dopamine$", "Measurement5", names(data))

#questions below are all missing
# names(data) <- gsub("^Q57_1$", "ConsequenceHb1", names(data))
# names(data) <- gsub("^Q57_2$", "ConsequenceHb2", names(data))
# names(data) <- gsub("^Q57_3$", "ConsequenceHb3", names(data))
# names(data) <- gsub("^Q57_4$", "ConsequenceHb4", names(data))
# names(data) <- gsub("^Q57_5$", "ConsequenceHb5", names(data))
# names(data) <- gsub("^Q57_6$", "ConsequenceHb6", names(data))
# names(data) <- gsub("^Q57_7$", "ConsequenceHb7", names(data))
# 
# names(data) <- gsub("^Q59_1$", "ConsequenceFer1", names(data))
# names(data) <- gsub("^Q59_2$", "ConsequenceFer2", names(data))
# names(data) <- gsub("^Q59_3$", "ConsequenceFer3", names(data))
# names(data) <- gsub("^Q59_4$", "ConsequenceFer4", names(data))
# names(data) <- gsub("^Q59_5$", "ConsequenceFer5", names(data))
# names(data) <- gsub("^Q59_6$", "ConsequenceFer6", names(data))
# names(data) <- gsub("^Q59_7$", "ConsequenceFer7", names(data))

names(data) <- gsub("^Have you ever been asked to take a break from donating blood because of low ferritin or hemoglobin\\?$", "Deferral", names(data))
names(data) <- gsub("^When was the last time you had a break from donating blood\\?$", "Last_Deferral", names(data))

names(data) <- gsub("^Understandable$", "DeferralInfo1", names(data))
names(data) <- gsub("^Informative$", "DeferralInfo2", names(data))
names(data) <- gsub("^Easy to remember$", "DeferralInfo3", names(data))
names(data) <- gsub("^Useful$", "DeferralInfo4", names(data))
names(data) <- gsub("^Clear$", "DeferralInfo5", names(data))

names(data) <- gsub("^Did you miss any information\\?$", "LackDeferralInfo_process", names(data))

#these have to be created from the column above
# names(data) <- gsub("^Q69_1$", "LackDeferralInfo1", names(data))
# names(data) <- gsub("^Q69_2$", "LackDeferralInfo2", names(data))
# names(data) <- gsub("^Q69_3$", "LackDeferralInfo3", names(data))
# names(data) <- gsub("^Q69_4$", "LackDeferralInfo4", names(data))
# names(data) <- gsub("^Q69_4_TEXT$", "LackDeferralInfoTEXT", names(data))

names(data) <- gsub("^Column$", "Deferral_Motivation1", names(data))
names(data) <- gsub("^The instruction 1$", "Deferral_Motivation2", names(data))

names(data) <- gsub("^Do you know which symptoms are common with iron deficiency\\? \\(multiple answers are possible\\)$", "SymptomsID_process", names(data))
#these have to be created from the one above
# names(data) <- gsub("^Q42_1$", "SymptomsID1", names(data))
# names(data) <- gsub("^Q42_2$", "SymptomsID2", names(data))
# names(data) <- gsub("^Q42_3$", "SymptomsID3", names(data))
# names(data) <- gsub("^Q42_4$", "SymptomsID4", names(data))
# names(data) <- gsub("^Q42_5$", "SymptomsID5", names(data))
# names(data) <- gsub("^Q42_6$", "SymptomsID6", names(data))
# names(data) <- gsub("^Q42_7$", "SymptomsID7", names(data))
# names(data) <- gsub("^Q42_8$", "SymptomsID8", names(data))
# names(data) <- gsub("^Q42_9$", "SymptomsID9", names(data))

names(data) <- gsub("^When you leave blood\\, you lose iron\\. Do you do anything before donating blood to avoid iron deficiency\\? \\(for example by eating an iron\\-rich diet or taking iron tablets\\)\\.$", "AccForIronLoss", names(data))
names(data) <- gsub("^How do you replace the iron loss that occurs when donating blood\\?$", "AccForIronLoss_process", names(data))
#these have to be created from the one above
# names(data) <- gsub("^Q51_1.0$", "AccForIronLoss1", names(data))
# names(data) <- gsub("^Q51_4.0$", "AccForIronLoss2", names(data))
# names(data) <- gsub("^Q51_2.0$", "AccForIronLoss3", names(data))
# names(data) <- gsub("^Q51_3.0$", "AccForIronLoss4A", names(data))
# names(data) <- gsub("^Q51_3_TEXT$", "AccForIronLoss4B", names(data))

names(data) <- gsub("^Blood donation staff$", "DeferralSource1", names(data))
#names(data) <- gsub("^Q53_2$", "DeferralSource2.NL", names(data))
names(data) <- gsub("^Healthcare personnel\\.\\.\\.85$", "DeferralSource3", names(data))
names(data) <- gsub("^The blood center\\'s website or brochure$", "DeferralSource4", names(data))
names(data) <- gsub("^Internet$", "DeferralSource5", names(data))
names(data) <- gsub("^Other source$", "DeferralSource6A", names(data))
#names(data) <- gsub("^Q88_1_TEXT$", "DeferralSource6B", names(data))

names(data) <- gsub("^Blood donation staff2$", "DeferralSource1_Confidence.FI", names(data))
#names(data) <- gsub("^Q53_2$", "DeferralSource2.NL", names(data))
names(data) <- gsub("^Healthcare personnel\\.\\.\\.90$", "DeferralSource3_Confidence.FI", names(data))
names(data) <- gsub("^Blood servicewebsite and\\/or brochure$", "DeferralSource4_Confidence.FI", names(data))
names(data) <- gsub("^Internet2$", "DeferralSource5_Confidence.FI", names(data))
names(data) <- gsub("^Other sources$", "DeferralSource6_Confidence.FI", names(data))
#names(data) <- gsub("^Q88_1_TEXT$", "DeferralSource6B", names(data))

names(data) <- gsub("^Bad \\(1\\) \\- Good \\(5\\)$", "Supplements1", names(data))
names(data) <- gsub("^Harmful \\(1\\) \\- Beneficial \\(5\\)$", "Supplements2", names(data))
names(data) <- gsub("^Should be advised against \\(1\\) \\- Should be recommended \\(5\\)$", "Supplements3", names(data))
names(data) <- gsub("^Meaningless \\(1\\) \\-Meaningful \\(5\\)$", "Supplements4", names(data))

names(data) <- gsub("^Do you take nutritional supplements\\? \\(e\\.g\\. vitamin C\\, vitamin D\\, calcium\\, iron\\, multivitamin tablets\\)$", "SupplementsUse", names(data))
names(data) <- gsub("^What is the name of the dietary supplement\\(s\\) you take or have taken\\?$", "SupplementsUse_Type_A", names(data))
# names(data) <- gsub("^Q73_2$", "SupplementsUse2A.NL", names(data))
# names(data) <- gsub("^Q73_3$", "SupplementsUse3A.NL", names(data))
# names(data) <- gsub("^Q73_1_TEXT$", "SupplementsUse1B.NL", names(data))
# names(data) <- gsub("^Q73_2_TEXT$", "SupplementsUse2B.NL", names(data))
# names(data) <- gsub("^Q73_3_TEXT$", "SupplementsUse3B.NL", names(data))

names(data) <- gsub("^I was recommended by the blood center or by healthcare professionals$", "SupplementsReason1", names(data))
names(data) <- gsub("^Due to health problems$", "SupplementsReason2", names(data))
names(data) <- gsub("^To improve my overall health$", "SupplementsReason3", names(data))
names(data) <- gsub("^Because of my diet \\(e\\.g\\. vegetarian diet\\)$", "SupplementsReason4", names(data))
names(data) <- gsub("^Because I leave blood$", "SupplementsReason5", names(data))
names(data) <- gsub("^Is there another reason you are taking supplements\\? \\(if so what\\)$", "SupplementsReason6A", names(data))
#names(data) <- gsub("^Q89_1_TEXT$", "SupplementsReason6B.NL", names(data))

names(data) <- gsub("^Can you imagine using nutritional supplements in the future\\? Under what conditions\\?$", "SupplementsOpen", names(data))
names(data) <- gsub("^Q34_1_TEXT$", "SupplementsOpen1", names(data))
names(data) <- gsub("^Q34_2_TEXT$", "SupplementsOpen2", names(data))

names(data) <- gsub("^Iron tablets are given by the blood center only to donors with low ferritin levels$", "IronSuppKnowledge1", names(data))
names(data) <- gsub("^Iron tablets can facilitate recovery after donating blood$", "IronSuppKnowledge2", names(data))
names(data) <- gsub("^Iron tablets can only be prescribed with a prescription$", "IronSuppKnowledge3", names(data))
names(data) <- gsub("^Iron tablets can cause side effects$", "IronSuppKnowledge4", names(data))

names(data) <- gsub("^Iron tablets are given by the blood center only to donors with low ferritin levels2$", "IronSuppKnowledge1_Certainty", names(data))
names(data) <- gsub("^Iron tablets can facilitate recovery after donating blood2$", "IronSuppKnowledge2_Certainty", names(data))
names(data) <- gsub("^Iron tablets can only be prescribed by prescription 2$", "IronSuppKnowledge3_Certainty", names(data))
names(data) <- gsub("^Iron tablets can cause side effects2$", "IronSuppKnowledge4_Certainty", names(data))

names(data) <- gsub("^Do you know the possible side effects of iron tablets\\? \\(multiple answers possible\\)$", "IronSuppSE_process", names(data))
#create from column above
# names(data) <- gsub("^Q37_1$", "IronSuppSE1", names(data))
# names(data) <- gsub("^Q37_2$", "IronSuppSE2", names(data))
# names(data) <- gsub("^Q37_3$", "IronSuppSE3", names(data))
# names(data) <- gsub("^Q37_4$", "IronSuppSE4", names(data))
# names(data) <- gsub("^Q37_5$", "IronSuppSE5", names(data))
# names(data) <- gsub("^Q37_6$", "IronSuppSE6", names(data))
# names(data) <- gsub("^Q37_7$", "IronSuppSE7.NL", names(data))
# names(data) <- gsub("^Q37_8$", "IronSuppSE8", names(data))
# names(data) <- gsub("^Q37_9$", "IronSuppSE9", names(data))

names(data) <- gsub("^Have you ever taken iron supplements\\?$", "IronSuppUse", names(data))

names(data) <- gsub("^Fitness$", "IronSuppEffect1", names(data))
names(data) <- gsub("^Fatigue3$", "IronSuppEffect2", names(data))
names(data) <- gsub("^Dizziness3$", "IronSuppEffect3", names(data))
names(data) <- gsub("^Headache3$", "IronSuppEffect4", names(data))
names(data) <- gsub("^Stomach problems$", "IronSuppEffect5", names(data))
names(data) <- gsub("^Do you find that iron supplements have any other health effects\\, if so what\\?$", "IronSuppEffect6A", names(data))
#names(data) <- gsub("^Q71_1_TEXT$", "IronSuppEffect6B", names(data))

names(data) <- gsub("^I am not sure if iron tablets have negative effects on my health$", "ConcernIronSupp1", names(data))
names(data) <- gsub("^I only want to take iron tablets if I have any problems$", "ConcernIronSupp2", names(data))
names(data) <- gsub("^I think iron tablets are making my health worse$", "ConcernIronSupp3", names(data))
names(data) <- gsub("^I\\'d rather eat more iron\\-containing food than take iron tablets$", "ConcernIronSupp4", names(data))
names(data) <- gsub("^Is there anything else that would prevent you from taking iron tablets\\? \\(Please specify what\\)$", "ConcernIronSupp5A", names(data))
#names(data) <- gsub("^If yes\\, please tell us the details\\.$", "ConcernIronSupp5B", names(data))

names(data) <- gsub("^Do you usually take the iron tablets you get from the blood center\\?$", "IronSuppSwe", names(data))

# names(data) <- gsub("^Q40_1$", "IronSuppSanq1", names(data))
# names(data) <- gsub("^Q40_2$", "IronSuppSanq2", names(data))
# names(data) <- gsub("^Q40_3$", "IronSuppSanq3", names(data))
# names(data) <- gsub("^Q40_4$", "IronSuppSanq4", names(data))
# names(data) <- gsub("^Q40_5$", "IronSuppSanq5", names(data))
# names(data) <- gsub("^Q40_6$", "IronSuppSanq6", names(data))
# names(data) <- gsub("^Q40_7$", "IronSuppSanq7", names(data))
# names(data) <- gsub("^Q40_8$", "IronSuppSanq8", names(data))
# names(data) <- gsub("^Q40_9$", "IronSuppSanq9", names(data))

# names(data) <- gsub("^Q41$", "IronSuppSanqReason", names(data)) #IronSupp_BloodService
# 
# names(data) <- gsub("^Q41_3_TEXT$", "IronSuppSanqReasonB", names(data))

names(data) <- gsub("^Suppose you are found to have low iron stores\\, which of the following options do you prefer\\? \\(rank from best \\(top\\) to worst option \\(bottom\\)$", "IronDefPref_process", names(data))
#have to be created from column above
# names(data) <- gsub("^Q77_1$", "IronDefPref1", names(data))
# names(data) <- gsub("^Q77_2$", "IronDefPref2", names(data))
# names(data) <- gsub("^Q77_3$", "IronDefPref3", names(data))
# names(data) <- gsub("^Q77_4$", "IronDefPref4", names(data))

names(data) <- gsub("^If I had health problems that could be remedied with the iron tablets$", "IronSuppSanqCondition1", names(data))
names(data) <- gsub("^To prevent potential health problems$", "IronSuppSanqCondition2", names(data))
names(data) <- gsub("^If it allowed me to continue donating blood regularly$", "IronSuppSanqCondition3", names(data))
names(data) <- gsub("^If the blood center doctor recommended me to take the iron tablets$", "IronSuppSanqCondition4", names(data))
names(data) <- gsub("^If there is scientific evidence that the iron tablets improve the body's iron stores$", "IronSuppSanqCondition5", names(data))
names(data) <- gsub("^I wouldn\\'t want to take the iron tablets$", "IronSuppSanqCondition6", names(data))

names(data) <- gsub("What do you think is the most important thing for the blood center to take into account regarding the formulation of recommendations for blood donors with low iron stores? (top - highest priority, along the bottom - lowest priority...", "IronSuppSanq_process", names(data), fixed=T)
#have to be created from column above
# names(data) <- gsub("^Q81_1$", "IronSuppSanqReason1", names(data))
# names(data) <- gsub("^Q81_2$", "IronSuppSanqReason2", names(data))
# names(data) <- gsub("^Q81_3$", "IronSuppSanqReason3", names(data))
# names(data) <- gsub("^Q81_4$", "IronSuppSanqReason4", names(data))
# names(data) <- gsub("^Q81_5$", "IronSuppSanqReason5", names(data))


colnames(data)
```
# Translate answers

```{r}
Data2<-data

#Age
data <- data %>%
  mutate(Age= as.numeric(Age),
         Age = cut(Age,
                            breaks = c(17, 20, 30, 40, 50, 60, 70),
                            labels = c("17-20", "21-30", "31-40", "41-50", "51-60", "61-70"),
                            right = TRUE, # Include the upper bound in the intervals
                            include.lowest = TRUE)) # Include the lowest value in the first interval

#create menstruation other
data <- data %>% mutate(Menstruation = ifelse(Sex == "MAN", NA, Menstruation), Menstruation_other = Menstruation, Menstruation_other = ifelse(Menstruation_other %in% c("Yes", "No, because of menopause", "No, because of contraception"), NA, Menstruation_other), Menstruation = ifelse(!(Menstruation %in% c("Yes", "No, because of menopause", "No, because of contraception", NA)), "No, other", Menstruation))

#recode Area:If you want to convert the responses Umeå is a city with more than 100 000 residents, whereas all other cities have less than 100 000 inhabitants.
data <- data %>% mutate(Area = ifelse(Area == "Umeå", "City, 100,000 or more", "City or town, < 100,000 residents"))

#Map education
data <- data %>% mutate(Education = ifelse(grepl("Gym", Education), "High school", Education), Education = ifelse(!(Education %in% c("Other",
"None",
"Primary school", 
"High school",
"College",
"Applied sciences",
"University")), "Other", Education))

#create lackdeferralinfo 
#LackDeferralInfo
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    LackDeferralInfo1 = sapply(as.character(LackDeferralInfo_process), function(x) ifelse(grepl("Information about what hemoglobin and\\/or ferritin are", x), TRUE, FALSE)),
    LackDeferralInfo5 = sapply(as.character(LackDeferralInfo_process), function(x) ifelse(grepl("Information about the consequences of low hemoglobin and\\/or ferritin", x), TRUE, FALSE)),
    LackDeferralInfo2 = sapply(as.character(LackDeferralInfo_process), function(x) ifelse(grepl("Information on how best to restore hemoglobin and\\/or ferritin levels", x), TRUE, FALSE)),
    LackDeferralInfo3 = sapply(as.character(LackDeferralInfo_process), function(x) ifelse(grepl("No information was missing", x), TRUE, FALSE)),
    
    # Capture any open text answers that do not match predefined options
    LackDeferralInfoTEXT = sapply(as.character(LackDeferralInfo_process), function(x) {
      if (is.na(x)) {
        return(NA)
      }
      predefined_patterns <- c("Information about what hemoglobin and\\/or ferritin are", 
                               "Information about the consequences of low hemoglobin and\\/or ferritin", 
                               "Information on how best to restore hemoglobin and\\/or ferritin levels", 
                               "No information was missing")
      # Remove predefined patterns from the string
      remaining_text <- trimws(gsub(paste(predefined_patterns, collapse = "|"), "", x))
      if (nchar(remaining_text) > 0) {
        return(remaining_text)
      } else {
        return(NA)
      }
    }),
    LackDeferralInfo5 = sapply(as.character(LackDeferralInfoTEXT), function(x) {
      if(is.na(x)){
      return(FALSE)
    } else if (nchar(x) > 0){
      return(TRUE)
    }
    }) 
  )

data <- data %>% select(-"LackDeferralInfo_process")

#create SymptomsID
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    SymptomsID1 = sapply(as.character(SymptomsID_process), function(x) ifelse(grepl("Restless", x), TRUE, FALSE)),
    SymptomsID2 = sapply(as.character(SymptomsID_process), function(x) ifelse(grepl("Dark stools", x), TRUE, FALSE)),
    SymptomsID3 = sapply(as.character(SymptomsID_process), function(x) ifelse(grepl("Fatigue", x), TRUE, FALSE)),
    SymptomsID4 = sapply(as.character(SymptomsID_process), function(x) ifelse(grepl("Stomach", x), TRUE, FALSE)),
    SymptomsID5 = sapply(as.character(SymptomsID_process), function(x) ifelse(grepl("chew", x), TRUE, FALSE)),
    SymptomsID6 = sapply(as.character(SymptomsID_process), function(x) ifelse(grepl("Headache", x), TRUE, FALSE)),
    SymptomsID7 = sapply(as.character(SymptomsID_process), function(x) ifelse(grepl("None of the above", x), TRUE, FALSE)),
    SymptomsID8 = sapply(as.character(SymptomsID_process), function(x) ifelse(grepl("I don\\'t know", x), TRUE, FALSE))
  )

data <- data %>% select(-"SymptomsID_process")

#create AccForIronLoss1-4
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    AccForIronLoss1 = sapply(as.character(AccForIronLoss_process), function(x) ifelse(grepl("diet", x), TRUE, FALSE)),
    AccForIronLoss2 = sapply(as.character(AccForIronLoss_process), function(x) ifelse(grepl("multivitamins", x), TRUE, FALSE)),
    AccForIronLoss3 = sapply(as.character(AccForIronLoss_process), function(x) ifelse(grepl("tablets", x), TRUE, FALSE)),
    AccForIronLoss5 = sapply(as.character(AccForIronLoss_process), function(x) ifelse(grepl("nothing", x), TRUE, FALSE))
  )

data <- data %>% select(-"AccForIronLoss_process")

#create supplements reason other
data <- data %>% mutate(SupplementsReason6B = ifelse(!is.na(SupplementsReason6A)&SupplementsReason6A!="No", SupplementsReason6A, NA), SupplementsReason6A = ifelse(!is.na(SupplementsReason6A)&SupplementsReason6A!="No", "Yes", SupplementsReason6A))

#transform supplements open into yes/no
data <- data %>% mutate(
    SupplementsOpen1 = ifelse(grepl("Yes", SupplementsOpen), SupplementsOpen, NA),
    SupplementsOpen2 = ifelse(grepl("No", SupplementsOpen), SupplementsOpen, NA),
    SupplementsOpen = ifelse(is.na(SupplementsOpen), NA, ifelse(grepl("Yes", SupplementsOpen), "Yes", "No"))
  )

#create ironsuppse1-9
data <- data %>%
  mutate(
    # Create binary columns for each predefined answer option using regex to match the content
    IronSuppSE1 = sapply(as.character(IronSuppSE_process), function(x) ifelse(grepl("Restless legs", x), TRUE, FALSE)),
    IronSuppSE2 = sapply(as.character(IronSuppSE_process), function(x) ifelse(grepl("Dark stools", x), TRUE, FALSE)),
    IronSuppSE3 = sapply(as.character(IronSuppSE_process), function(x) ifelse(grepl("Fatigue", x), TRUE, FALSE)),
    IronSuppSE4 = sapply(as.character(IronSuppSE_process), function(x) ifelse(grepl("Stomach upset", x), TRUE, FALSE)),
    IronSuppSE5 = sapply(as.character(IronSuppSE_process), function(x) ifelse(grepl("Urge to chew on non\\-edible products like chalks\\, ice cubes\\, pen caps\\, or raw pasta", x), TRUE, FALSE)),
    IronSuppSE6 = sapply(as.character(IronSuppSE_process), function(x) ifelse(grepl("Headaches", x), TRUE, FALSE)),
    IronSuppSE7 = sapply(as.character(IronSuppSE_process), function(x) ifelse(grepl("All of the above", x), TRUE, FALSE)),
    IronSuppSE9 = sapply(as.character(IronSuppSE_process), function(x) ifelse(grepl("I don\\'t know", x), TRUE, FALSE))
  )

data <- data %>% select(-"IronSuppSE_process")

#transform IronSuppEffect6A
data <- data %>% select(-"IronSuppEffect6A") #removing this because we have only the text, not whether it increased/decreased

#transform ConcernIronSupp5A
data <- data %>% mutate(ConcernIronSupp5B = ifelse(!is.na(ConcernIronSupp5A)&ConcernIronSupp5A!="No", ConcernIronSupp5A, NA), ConcernIronSupp5A = ifelse(!is.na(ConcernIronSupp5A)&ConcernIronSupp5A!="No", "Yes", ConcernIronSupp5A))

#create irondefpref1-5
split_data <- strsplit(data$IronDefPref_process, ";")


options <- c("A longer interval between future blood donations",
             "Dietary recommendations that potentially allow me to donate blood at the same frequency as before (would require follow-up with additional ferritin checks)",
             "Obtain more iron tablets that allow unchanged blood donation frequency",
             "Stop donating blood",
             "Give as before despite the low iron stores")
options <- c("longer interval",
             "Dietary",
             "tablets",
             "Stop",
             "despite")

# Convert the list into a data frame with each component of the split as a column
split_df <- do.call(rbind, lapply(split_data, function(x) {
  length(x) <- 5 # Ensure that there are exactly 5 elements in each row (pad with NA if necessary)
  return(x)
}))

colnames(split_df) <- paste0("Rank", 1:5)

# Assign column names
for (i in seq_along(options)) {
  column_name <- paste0("IronDefPref", i)
  data[[column_name]] <- NA
}

# Combine the existing data with split_df (assuming split_df has Rank1 to Rank5)
data <- cbind(data, split_df)

# Loop through the options to fill in the ranks using grepl
for (i in seq_along(options)) {
  option <- options[i]
  column_name <- paste0("IronDefPref", i)
  
  # Apply grepl to find partial matches for each row and assign the corresponding rank
  data[[column_name]] <- apply(data[, paste0("Rank", 1:5)], 1, function(x) {
    # Use grepl to check for partial matches in each Rank column
    match_idx <- which(sapply(x, function(rank_value) grepl(option, rank_value)))
    
    # Return the position of the first match or NA if none
    if (length(match_idx) > 0) {
      return(match_idx[1])  # Return the first match index (Rank1, Rank2, etc.)
    } else {
      return(NA)  # If no match, return NA
    }
  })
}

# Remove unnecessary columns after processing
data <- data %>% select(-c(Rank1, Rank2, Rank3, Rank4, Rank5, IronDefPref_process))

#IronSuppSanq_process
split_data <- strsplit(data$IronSuppSanq_process, ";")


options <- c("continue ",
             "other",
             "healthy",
             "safe",
             "never",
             "effectively")

# Convert the list into a data frame with each component of the split as a column
split_df <- do.call(rbind, lapply(split_data, function(x) {
  length(x) <- 6 # Ensure that there are exactly 5 elements in each row (pad with NA if necessary)
  return(x)
}))

colnames(split_df) <- paste0("Rank", 1:6)

# Assign column names
for (i in seq_along(options)) {
  column_name <- paste0("IronSuppSanqReason", i)
  data[[column_name]] <- NA
}

# Combine the existing data with split_df (assuming split_df has Rank1 to Rank5)
data <- cbind(data, split_df)

# Loop through the options to fill in the ranks using grepl
for (i in seq_along(options)) {
  option <- options[i]
  column_name <- paste0("IronSuppSanqReason", i)
  
  # Apply grepl to find partial matches for each row and assign the corresponding rank
  data[[column_name]] <- apply(data[, paste0("Rank", 1:6)], 1, function(x) {
    # Use grepl to check for partial matches in each Rank column
    match_idx <- which(sapply(x, function(rank_value) grepl(option, rank_value)))
    
    # Return the position of the first match or NA if none
    if (length(match_idx) > 0) {
      return(match_idx[1])  # Return the first match index (Rank1, Rank2, etc.)
    } else {
      return(NA)  # If no match, return NA
    }
  })
}

data <- data %>% select(-c(Rank1, Rank2, Rank3, Rank4, Rank5, Rank6,IronSuppSanq_process, IronSuppSanqReason3)) %>% rename(IronSuppSanqReason4.FI = IronSuppSanqReason4, IronSuppSanqReason6.FI = IronSuppSanqReason6)

#Deferral_Motivation
data <- data %>%
  mutate(Deferral_Motivation1 = case_when(
    Deferral_Motivation1 == "Decreased my willingness to give blood" ~ 1,
    Deferral_Motivation1 == "Slightly decreased my willingness to give blood" ~ 2,
    Deferral_Motivation1 == "No effect" ~ 3,
    Deferral_Motivation1 == "Made me slightly more inclined to leave blood" ~ 4,
    Deferral_Motivation1 == "Made me distinctly more inclined to leave blood" ~ 5,
    TRUE ~ NA_real_  # Handle any unexpected values
  ))

data <- data %>%
  mutate(Deferral_Motivation2 = case_when(
    Deferral_Motivation2 == "Reduces my desire to give blood" ~ 1,
    Deferral_Motivation2 == "Slightly decreased my willingness to give blood" ~ 2,
    Deferral_Motivation2 == "No effect" ~ 3,
    Deferral_Motivation2 == "Made me slightly more inclined to leave blood" ~ 4,
    Deferral_Motivation2 == "Made me distinctly more inclined to leave blood" ~ 5,
    TRUE ~ NA_real_  # Handle any unexpected values
  ))



#replace yes and no with true and false
data <- data %>%
  mutate(across(matches("^SupplementsReason[0-8]$"),
                ~ case_when(
                    . == "Yes" ~ TRUE,
                    . == "No"  ~ FALSE  # Leave other values unchanged if any
                ))) %>%
  mutate(across(matches("^DeferralSource([0-8]|6A)$"),
                ~ case_when(
                    . == "Yes" ~ TRUE,
                    . == "No"  ~ FALSE  # Leave other values unchanged if any
                )))
#Ironsuppknowledge_certainty
data <- data %>%
  # Apply transformations to the columns matching the regex
  mutate(across(matches("^Ironsuppknowledge[0-8]_Certainty$"), 
                ~ case_when(
                    . == "Sure" ~ "Sure",
                    . == "Uncertain"  ~ "Not sure"  # Leave other values unchanged if any
                )))

```

## Mapping functions 

```{r}
#what it becomes = what is it

#col 11 and 12 

f1 <- function (x){
  x <- fct_recode(x,  
                  "Disagree" = "Mostly disagree",
                  "Neutral"="Partially agree",
                  "Agree"="Mostly agree",
                  "Totally disagree" =  "Do not agree at all",
                  "Totally agree" = "Totally agree"
  )
  x <- factor(
    x,
    levels = c("Totally disagree","Disagree","Neutral","Agree","Totally agree"),
    ordered = TRUE
  )
  x
}

f2 <- function (x){
  x <- fct_recode(x,  
                  "Never" = "Never",
                  "Sometimes" = "Sometimes",
                  "Often" = "Often",
                  "Always" = "Always"
  )
  x <- factor(
    x,
    levels = c("Never","Sometimes","Often","Always"),
    ordered = TRUE
  )
  x  
}

f3 <- function (x){
  x <- fct_recode(x,  
                  "Yes" = "1",
                  "No" = "2",
                  "I don't know" = "4"
                  )
  x <- factor(
    x,
    levels=c("Yes","No","I don't know"),
    ordered = TRUE
  )
  x
}

f4 <- function (x){
  x <- fct_recode(x,
                  "FALSE" = "False",
                  "TRUE" = "True"
                  
                  )
  x <- factor(
    x,
    levels=c("FALSE","TRUE"),
  )
  x
}

f5 <- function (x){
  x <- fct_recode(x,
                  "50/50" = "Not sure",
                  "100%" = "Sure"
                  
                  )
  x <- factor(
    x,
    levels=c("50/50","100%"),
  )
  x
}


f6 <- function (x){
  x <- fct_recode(x,
                  "No" = "No",
                  "Yes" = "Yes"
                  
                  )
  x <- factor(
    x,
    levels=c("No","Yes"),
  )
  x
}


f7 <- function(x) {
  x <- as.character(x)
  x[is.na(x)] <- "FALSE"  # Convert NA values to "FALSE"
  x <- forcats::fct_recode(x,
                           "FALSE" = NA_character_,
                           "TRUE" = "1"
  )
  x <- factor(x,
              levels = c("FALSE", "TRUE")
  )
  x
}

f8 <- function (x){ #HIER AANPASSEN
  x <- fct_recode(x,  
                  "Increased" = "Increases",
                  "No effect" = "No effect",
                  "Decreased" = "Reduces",
                  "Don't know" = "Don't know"
  )
  x <- factor(
    x,
    levels=c("Increased","No effect","Decreased", "Don't know"),
  )
  x
}

f9 <- function (x){
  x <- fct_recode(x,
                  "Agree" = "Agree",
                  "Disagree" = "Disagree"
                  )
  x <- factor(
    x,
    levels=c("Agree","Disagree"),
  )
  x
}

f10 <- function (x){
  x <- factor(
    x,
    levels=c("1","2","3","4","5","6"),
    ordered = TRUE
  )
  x
}


f11 <- function (x){
  x <- factor(
    x,
    levels=c("1","2","3","4","5"),
    ordered = TRUE
  )
  x
}

f12 <- function (x){
  x <- fct_recode(x,
                  "1" = "Never experienced",
                  "2" = "Very mild",
                  "3" = "Mild",
                  "4" = "Moderate",
                  "5" = "Difficult/tangible"
                  )
  x <- factor(
    x,
    levels=c("1","2","3","4","5"),
  )
  x
}

f13 <- function (x){
  x <- fct_recode(x,
                  "1" = "Not at all",
                  "2" = "Almost not",
                  "3" = "To some extent",
                  "4" = "Mostly",
                  "5" = "Completely"
                  )
  x <- factor(
    x,
    levels=c("1","2","3","4","5"),
  )
  x
}


f14 <- function (x){
  x <- fct_recode(x,
                  "1" = "Not at all",
                  "2" = "A little",
                  "3" = "Means",
                  "4" = "Mostly",
                  "5" = "Completely"
                  )
  x <- factor(
    x,
    levels=c("1","2","3","4","5"),
  )
  x
}
```


## Main processing



```{r}
#Columns containing just 0 or 1
# Columns containing just yes and no (and NA)
yesandnocols <- data %>%
  summarise(across(everything(), ~ all(.x %in% c("Yes", "No", NA), na.rm = TRUE)))

#Columns containing always, often, sometimes and never
aosncols <- data %>%
  summarise(across(everything(), ~ all(.x %in% c("Always", "Often", "Sometimes", "Never", NA), na.rm = TRUE)))

#Columns containing just 0 or 0 & 1 
zero1cols <- data %>%  
  summarise(across(everything(), ~ {
    unique_vals <- sort(unique(.x))
    all(unique_vals == 0) || all(unique_vals == c(0, 1) || all(unique_vals==1))
  }))

FTcols <- data %>%  
  summarise(across(everything(), ~ {
    unique_vals <- sort(unique(.x))
    all(unique_vals == "False") || all(unique_vals == c("False", "True") || all(unique_vals=="True"))
  }))

#tmp[] <- lapply(tmp, as.factor)

cols12345 <- c("Deferral_Motivation1", "Deferral_Motivation2", "Supplements1", "Supplements2", "Supplements3", "Supplements4", "IronDefPref1",  "IronDefPref2", "IronDefPref3", "IronDefPref4", "IronDefPref5", "IronSuppSanqReason1", "IronSuppSanqReason2", "IronSuppSanqReason5",  "IronSuppSanqReason4.FI",  "IronSuppSanqReason6.FI")

cols16 <- c(
  ) # i think in sweden these are also 1-5 because one option was msising

# Statements: Totally disagree - Totally agree
tmp <- data %>% 
  mutate_at(vars(which(str_detect(colnames(data) ,"OpinionBloodService"))), ~ f1(.) ) %>%  
  mutate_at(vars(which(as.logical(aosncols))), ~ f2(.)) %>% #symptoms columns
  #mutate_at(vars(which(str_detect(colnames(data), "Measurement"))), ~ f3(.))%>% 
  mutate_at(vars(which(as.logical(FTcols))), ~ f4(.)) %>%
  mutate_at(vars(which(str_detect(colnames(data), "_Certainty"))), ~ f5(.)) %>% 
  mutate_at(vars(which(as.logical(yesandnocols))), ~ f6(.) ) %>%
  mutate_at(vars(which(as.logical(zero1cols))), ~ f7(.) ) %>% 
  mutate_at(vars(which(str_detect(colnames(data), "\\bIronSuppEffect2\\b|\\bIronSuppEffect3\\b|\\bIronSuppEffect4\\b|\\bIronSuppEffect5\\b|\\bIronSuppEffect6A\\b"))), ~ f8(.))  %>%
  mutate_at(vars(which(str_detect(colnames(data) ,"IronSuppSanqCondition"))), ~ f9(.) ) %>%
  mutate_at(vars(match(cols12345, colnames(data))), ~ f11(.)) %>%
  mutate_at(vars(which(str_detect(colnames(data), "_Severity"))), ~ f12(.)) %>%
  mutate_at(vars(which(str_detect(colnames(data), "\\bDeferralInfo1\\b|\\bDeferralInfo2\\b|\\bDeferralInfo3\\b|\\bDeferralInfo4\\b|\\bDeferralInfo5\\b"))), ~ f13(.))  %>%
  mutate_at(vars(which(str_detect(colnames(data), "_Confidence.FI"))), ~ f14(.))


tmp <- tmp %>% 
  mutate(
    Sex =  fct_recode(Sex,  
                  "Male" = "MAN",
                  "Female" = "Woman",
                  "Non-binary" = "Non-binary",
                  "Other" = "Other",
                  "Prefer not to say" = "Prefer not to say"
                  ),
    Menstruation = fct_recode(Menstruation,  
    "Yes" = "Yes",
    "No, due to current pregnancy" = "No, because of pregnancy",
    "No, due to menopause" = "No, because of menopause",
    "No, due to current contraception use" = "No, because of contraception",
    "No, due to a different reason" = "No, other"
     ),  
    Area = fct_recode(Area,
                     "In a city of over 100,000 people" = "City, 100,000 or more",
                    "In a town of less than 100,000 people, or in the countryside" = "City or town, < 100,000 residents"
                                                 
     ),
    Education =fct_recode(Education,
"None" = "None",
"Primary school" = "Primary school", 
"High school" = "High school",
"College" = "College ",
"Applied sciences" = "Applied sciences",
"University" = "University",
"Other" = "Other"
      ) %>% fct_relevel(
"Other",
"None",
"Primary school", 
"High school",
"College",
"Applied sciences",
"University"
      ) %>% as.ordered(.)
,
   Health = fct_recode(Health,
"Excellent" =  "Excellent",
"Very good" =  "Very good",
"Good" =   "Good",
"Fair" =  "Less good",
"Poor" =  "Poor"
) %>% fct_relevel(.,
"Poor",
"Fair",
"Good",
"Very good",
"Excellent",  
) %>% as.ordered(.)
,
    Diet =as.factor(Diet)%>%
    fct_recode(.,
      "Yes, vegan (no animal products)" =  "Yes, vegan (no animal products)",
"Yes, vegetarian (no meat and no fish)" =  "Yes, vegetarian (no meat or fish)",
"Yes, pescatarian (no meat)" =  "Yes, pescatarian (no meat)",
"Yes, flexitarian (at least three days without meat)" = "Yes, flexitarian (at least 3 days per week without meat)" ,
"No" = "No"
    ),    
    Number_donation = fct_recode(Number_donation,
      "1 - 3 times" = "1-3 times",
"4 - 10 times" = "4-10 times",
"More than 10 times" = "More than 10 times",
"None" = "4",
"I don't know" = "5",
"More than 30 times" = "6"
    ) %>% fct_relevel(.,
    "I don't know",
    "None",
    "1 - 3 times",
"4 - 10 times",
"More than 10 times",
"More than 30 times"
) %>% as.ordered(.)
,
    Last_donation = fct_recode(Last_donation,
"Within the last 6 months" =  "Less than 6 months ago",
"6 months to a year ago" =  "6-12 months ago",
"1-2 years ago" =  "13-24 months ago",
"More than 2 years ago" =  "More than 24 months ago",
"I cannot remember" = "5",
    ) %>% fct_relevel(.,
"Within the last 6 months",
"6 months to a year ago",
"1-2 years ago",
"More than 2 years ago",
"I cannot remember"
                      ) %>% as.ordered(.)

,
Deferral
=fct_recode(Deferral,
            "No" = "No",
            "I don’t know"= "Don't know",
            "Yes, once "= "Yes, once",
            "Yes, multiple times "= "Yes, several times"
            
),
Last_Deferral = fct_recode(Last_Deferral,
"Currently" = "I am currently deferred",
"Within the last 6 months" =  "Within the last 6 months",
"6 months to a year ago" =  "6-12 months ago",
"1-2 years ago" =  "13-24 months ago",
"More than 2 years ago" = "More than 24 months ago",
"I cannot remember" = "Can't remember",
    ) %>% fct_relevel(.,
                      "Currently",
"Within the last 6 months",
"6 months to a year ago",
"1-2 years ago",
"More than 2 years ago",
"I cannot remember"
                      ) %>% as.ordered(.),
       
SupplementsUse.NL =
  fct_recode(SupplementsUse,
             "Never" = "Never",
             "Not anymore" =  "Not now, but before",
             "Sometimes" =  "Sometimes",
             "Every day" =  "Every day"
  ),
IronSuppUse =
  fct_recode(IronSuppUse,
             "No, I have never taken iron (containing) supplements" = "No, I have never taken iron supplements",
             "Yes, I am currently taking iron supplements ordered to me by a doctor"    = "Yes, I have an ongoing treatment that I received from the blood center",
             "Yes, I am currently taking iron supplements from a drugstore"   =  "Yes, I have ongoing treatment that I bought without a prescription at a pharmacy", 
             "Yes, I take multivitamins that contain iron"  =  "Yes, I take multivitamins that contain iron",
             "I have taken iron supplements/multivitamins before, but not anymore"      = "I used to take iron/multivitamin tablets, but no longer do" 
             
  )



  )
  
```

# remove the open text answers for making the summary
```{r}
#strings <- data %>% select(where(is.character))

tmp <- tmp %>%
  select(-c("ID","SupplementsUse_Type_A","LackDeferralInfoTEXT", "SupplementsReason6B", "SupplementsOpen1", "SupplementsOpen2", "Menstruation_other", "ConcernIronSupp5B"))
```

# Export summary data

```{r}
#Export summary count data
DataExport <- apply(tmp,MARGIN = 2,FUN = table)
saveRDS(DataExport,file="L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/2. Datasets/individual/1. iDPS - Summary data - SWE.RDS")

```


```{r}
# dfSummary(data,
#           plain.ascii = FALSE, 
#           style = "grid", 
#           graph.col = TRUE,
#           graph.magnif = 0.75, 
#           valid.col = FALSE, 
#           tmp.img.dir = "/tmp")


dataswe <- saveRDS(tmp, file = "L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/A Source documents/iDPS - SWE/dataswe.RDS")

```
