==> src/1_translate_and_describe_USA.Rmd <==
---
title: "Donor iron supplement perception study data description and translation - USA"
author: "Amber Meulenbeld"
date: "`r Sys.time()`"
output:
  html_document:
    df_print: paged
  pdf_document:
    fig_caption: yes
    number_sections: yes
---

\tableofcontents


```{r setup, include=FALSE}

  
library(knitr)
opts_chunk$set(#results = 'asis',      # Can also be set at the chunk-level. IT IS NEEDED IN THE dfsummary chunck
               comment = NA,
               prompt  = FALSE,
               cache   = FALSE,
               echo = TRUE,
               message = FALSE,
               warning = FALSE,
               width.cutoff=60
)
library(summarytools)
st_options(plain.ascii = FALSE,        # Always use this option in Rmd documents
           style        = "rmarkdown", # Always use this option in Rmd documents
           footnote     = NA,          # Makes html-rendered results more concise
           subtitle.emphasis = FALSE)  # Improves layout with some rmardown themes
library(tidyverse)
library(readxl)
library(lubridate)
library(summarytools)

```


# Read and summarise data


```{r}
Data <- read_excel("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/A Source documents/iDPS - USA/Iron_Supplementation_Practices_Survey_20240930.xlsx")

data <- Data
#colnames(data)
```

# Change column names to unify

```{r}
#most correct colnames are already in 
names(data) <- gsub("^Number_donation_rbc$", "Number_donation", names(data))
names(data) <- gsub("^Last_donation_rbc$", "Last_donation", names(data))

names(data) <- gsub("^Last_Deferral$", "Last_Deferral_hb", names(data))
names(data) <- gsub("^Deferral$", "Deferral_hb", names(data))

names(data) <- gsub("^Knowledge_hb1$", "KnowledgeHb1", names(data))
names(data) <- gsub("^Knowledge_hb2$", "KnowledgeHb2", names(data))
names(data) <- gsub("^Knowledge_hb3$", "KnowledgeHb3", names(data))
names(data) <- gsub("^Knowledge_hb4$", "KnowledgeHb4", names(data))
names(data) <- gsub("^Knowledge_hb1_Conf$", "KnowledgeHb1_Certainty", names(data))
names(data) <- gsub("^Knowledge_hb2_Conf$", "KnowledgeHb2_Certainty", names(data))
names(data) <- gsub("^Knowledge_hb3_Conf$", "KnowledgeHb3_Certainty", names(data))
names(data) <- gsub("^Knowledge_hb4_Conf$", "KnowledgeHb4_Certainty", names(data))

names(data) <- gsub("^Knowledge_Fer1$", "KnowledgeFer1", names(data))
names(data) <- gsub("^Knowledge_Fer2$", "KnowledgeFer2", names(data))
names(data) <- gsub("^Knowledge_Fer3$", "KnowledgeFer3", names(data))
names(data) <- gsub("^Knowledge_Fer4$", "KnowledgeFer4", names(data))
names(data) <- gsub("^Knowledge_Fer1_Conf$", "KnowledgeFer1_Certainty", names(data))
names(data) <- gsub("^Knowledge_Fer2_Conf$", "KnowledgeFer2_Certainty", names(data))
names(data) <- gsub("^Knowledge_Fer3_Conf$", "KnowledgeFer3_Certainty", names(data))
names(data) <- gsub("^Knowledge_Fer4_Conf$", "KnowledgeFer4_Certainty", names(data))

names(data) <- gsub("^Knowledge1_Conf$", "Knowledge1_Certainty", names(data))
names(data) <- gsub("^Knowledge2_Conf$", "Knowledge2_Certainty", names(data))
names(data) <- gsub("^Knowledge3_Conf$", "Knowledge3_Certainty", names(data))
names(data) <- gsub("^Knowledge4_Conf$", "Knowledge4_Certainty", names(data))

names(data) <- gsub("^DeferralInfo1$", "DeferralInfo_hb1", names(data))
names(data) <- gsub("^DeferralInfo2$", "DeferralInfo_hb2", names(data))
names(data) <- gsub("^DeferralInfo3$", "DeferralInfo_hb3", names(data))
names(data) <- gsub("^DeferralInfo4$", "DeferralInfo_hb4", names(data))

names(data) <- gsub("^LackDeferralInfo1$", "LackDeferralInfo_hb1", names(data))
names(data) <- gsub("^LackDeferralInfo2$", "LackDeferralInfo_hb2", names(data))
names(data) <- gsub("^LackDeferralInfo3$", "LackDeferralInfo_hb3", names(data))
names(data) <- gsub("^LackDeferralInfo4$", "LackDeferralInfo_hb4", names(data))

names(data) <- gsub("^DeferralMotivation1$", "DeferralMotivation_hb1", names(data))
names(data) <- gsub("^DeferralMotivation2$", "DeferralMotivation_hb2", names(data))

names(data) <- gsub("^DeferralSourceTrust1$", "DeferralSource1_Confidence.FI", names(data))
names(data) <- gsub("^DeferralSourceTrust3$", "DeferralSource3_Confidence.FI", names(data))
names(data) <- gsub("^DeferralSourceTrust4$", "DeferralSource4_Confidence.FI", names(data))
names(data) <- gsub("^DeferralSourceTrust5$", "DeferralSource5_Confidence.FI", names(data))
names(data) <- gsub("^DeferralSourceTrust6_ARC$", "DeferralSource6_ARC_Confidence.FI", names(data))
names(data) <- gsub("^DeferralSourceTrust6$", "DeferralSource6_Confidence.FI", names(data))

names(data) <- gsub("^IronSuppKnowledge_Conf1$", "IronSuppKnowledge1_Certainty", names(data))
names(data) <- gsub("^IronSuppKnowledge_Conf2$", "IronSuppKnowledge2_Certainty", names(data))
names(data) <- gsub("^IronSuppKnowledge_Conf3$", "IronSuppKnowledger3_Certainty", names(data))
names(data) <- gsub("^IronSuppKnowledge_Conf4$", "IronSuppKnowledge4_Certainty", names(data))

names(data) <- gsub("^AccForIronLoss_before$", "AccForIronLoss", names(data))
names(data) <- gsub("^AccForIronLoss_before1$", "AccForIronLoss1", names(data))
names(data) <- gsub("^AccForIronLoss_before2$", "AccForIronLoss2", names(data))
names(data) <- gsub("^AccForIronLoss_before3$", "AccForIronLoss3", names(data))
names(data) <- gsub("^AccForIronLoss_before4$", "AccForIronLoss4A", names(data))

names(data) <- gsub("^SupplementsUse$", "SupplementsUse.NL", names(data))

names(data) <- gsub("^SupplementsUse_Fe1$", "IronSuppUse1", names(data))
names(data) <- gsub("^SupplementsUse_Fe2$", "IronSuppUse2", names(data))
names(data) <- gsub("^SupplementsUse_Fe3$", "IronSuppUse3", names(data))
names(data) <- gsub("^SupplementsUse_Fe4$", "IronSuppUse4", names(data))
names(data) <- gsub("^SupplementsUse_Fe5$", "IronSuppUse5", names(data))
names(data) <- gsub("^SupplementsUse_Fe6$", "IronSuppUse6", names(data))
names(data) <- gsub("^SupplementsUse_Fe7$", "IronSuppUse7", names(data))

names(data) <- gsub("^IronSuppEffect6$", "IronSuppEffect6A", names(data))

names(data) <- gsub("^IronDefPref4$", "IronDefPref5", names(data))
```
# Remove columns

```{r}
# remove the supplements columns
regex_pattern <- "^SupplementsUse_(A|Mg|CurcTurm|MV[1-6]|B|Zn|Mel|C|Clgn|Fibr|Mela|D|Prob|Qrctn|K|Omega3|Protein|Ca|CoQ10|GrnT)(_Motiv[1-7])?$"

# Exclude columns that match the regex pattern
data <- data %>% select(-matches(regex_pattern))
data <- data %>% select(-c(IronSuppEffect6A))

colnames(data)
```

# Transform column

```{r}
#Accforironloss
data <- data %>%
  mutate_at(
    vars(which(str_detect(colnames(data), "^AccForIronLoss([0-9]|4A)$"))),
    ~ case_when(
        . %in% c(1, 2, 3, 4) ~ 1,   
        AccForIronLoss == 1 & is.na(.) ~ 2,                  # If AccForIronLoss is 1 but column value is not 1, set to 2
        TRUE ~ as.numeric(as.character(.))                   
    )
  )


#samenvoegen Last_deferral_hb en last_deferral_lowFer
data$Last_Deferral <- pmin(data$Last_Deferral_hb, data$Last_Deferral_lowFer, na.rm = TRUE)

#samenvoegen deferral_hb en deferral_lowFer
data$Deferral <- apply(data[, c("Deferral_hb", "Deferral_lowFer")], 1, function(row) {
  if (2 %in% row) {
    return(2)
  } else if (1 %in% row) {
    return(1)
  } else if (3 %in% row) {
    return(3)
  } else if (4 %in% row){
    return(4)
  } else { 
    return(NA)
    }
})

#ConsequenceHb all numbers to 1
data <- data %>%
  mutate_at(
    vars(which(str_detect(colnames(data), "^ConsequenceHb[0-9](_ARC)?$"))),
    ~ case_when(
        . %in% c(1, 2, 3, 4,5,6,7,8) ~ 1,   
        is.na(.) ~ 0)                            
    )

#ConsequenceFer all numbers to 1
data <- data %>%
  mutate_at(
    vars(which(str_detect(colnames(data), "^ConsequenceFer[0-9](_ARC)?$"))),
    ~ case_when(
        . %in% c(1, 2, 3, 4,5,6,7) ~ 1,   
        is.na(.) ~ 0)                          
    )

#LackDeferralInfo all numbers to 1
data <- data %>%
  mutate_at(
    vars(which(str_detect(colnames(data), "^LackDeferralInfo_hb[0-9]$"))),
    ~ case_when(
        . %in% c(1, 2, 3, 4,5) ~ 1,    
        is.na(.) ~ 0)                             
    )

#LackDeferralInfo_lowfer all numbers to 1
data <- data %>%
  mutate_at(
    vars(which(str_detect(colnames(data), "^LackDeferralInfo_lowFer[0-9]$"))),
    ~ case_when(
        . %in% c(1, 2, 3, 4,5) ~ 1,  
        is.na(.) ~ 0)                            
    )
#SymptomsID all numbers to 1
data <- data %>%
  mutate_at(
    vars(which(str_detect(colnames(data), "^SymptomsID[0-9]$"))),
    ~ case_when(
        . %in% c(1, 2, 3, 4,5,6,7,8) ~ 1,   
        is.na(.) ~ 0)      
  )

#IronSuppUse all numbers to 1 
data <- data %>%
  mutate_at(
    vars(which(str_detect(colnames(data), "^IronSuppUse[0-9]$"))),
    ~ case_when(
        . %in% c(1, 2, 3, 4,5,6,7) ~ 1,  
        is.na(.) ~ 0)                          
    )

#SUPPLEMENTSUSE_FE_REASON all numbers to 1 (index column for No is any SupplementsUse_Fe except 7)
data <- data %>%
  mutate_at(
    vars(which(str_detect(colnames(data), "^SupplementsUse_Fe_Reason[0-9]$"))),
    ~ case_when(
        . %in% c(1, 2, 3, 4,5,6) ~ 1,   
        (IronSuppUse1 == 1|IronSuppUse2==1|IronSuppUse3==1|IronSuppUse4==1|IronSuppUse5==1) & is.na(.) ~ 2,                  # If any of IronSuppUse[1-5] is 1 but column value is NA, set to 2
        TRUE ~ as.numeric(as.character(.))                   
    )
  )


#IRONSUPPSE all numbers to 1
data <- data %>%
  mutate_at(
    vars(which(str_detect(colnames(data), "^IronSuppSE[0-9]$"))),
    ~ case_when(
        . %in% c(1, 2, 3, 4,5,6,7,8,9) ~ 1,  
        is.na(.) ~ 0)                          
    )

```


## Mapping functions 

```{r}

f1 <- function (x){
  x <- fct_recode(x,  
                  "Disagree" = "2",
                  "Neutral"="3",
                  "Agree"="4",
                  "Totally disagree" =  "1",
                  "Totally agree" = "5"
  )
  x <- factor(
    x,
    levels = c("Totally disagree","Disagree","Neutral","Agree","Totally agree"),
    ordered = TRUE
  )
  x
}

f2 <- function (x){
  x <- fct_recode(x,  
                  "Never" = "1",
                  "Sometimes" = "2",
                  "Often" = "3",
                  "Always" = "4"
  )
  x <- factor(
    x,
    levels = c("Never","Sometimes","Often","Always"),
    ordered = TRUE
  )
  x  
}

f3 <- function (x){
  x <- fct_recode(x,  
                  "Yes" = "1",
                  "No" = "2",
                  "I don't know" = "3"
                  )
  x <- factor(
    x,
    levels=c("Yes","No","I don't know"),
    ordered = TRUE
  )
  x
}

f4 <- function (x){
  x <- fct_recode(x,
                  "FALSE" = "2",
                  "TRUE" = "1"
                  
                  )
  x <- factor(
    x,
    levels=c("FALSE","TRUE"),
  )
  x
}

f5 <- function (x){
  x <- fct_recode(x,
                  "50/50" = "1",
                  "100%" = "2"
                  
                  )
  x <- factor(
    x,
    levels=c("50/50","100%"),
  )
  x
}


f6 <- function (x){
  x <- fct_recode(x,
                  "No" = "2",
                  "Yes" = "1"
                  
                  )
  x <- factor(
    x,
    levels=c("No","Yes"),
  )
  x
}


f7 <- function(x) {
  x <- as.character(x)
  x[is.na(x)] <- "FALSE"  # Convert NA values to "FALSE"
  x <- forcats::fct_recode(x,
                           "FALSE" = "0",
                           "TRUE" = "1"
  )
  x <- factor(x,
              levels = c("FALSE", "TRUE")
  )
  x
}

f8 <- function (x){ #HIER AANPASSEN
  x <- fct_recode(x,  
                  "Increased" = "3",
                  "No effect" = "2",
                  "Decreased" = "1"
  )
  x <- factor(
    x,
    levels=c("Increased","No effect","Decreased"),
  )
  x
}

f9 <- function (x){
  x <- fct_recode(x,
                  "Agree" = "1",
                  "Disagree" = "2"
                  )
  x <- factor(
    x,
    levels=c("Agree","Disagree"),
  )
  x
}


f11 <- function (x){
  x <- factor(
    x,
    levels=c("1","2","3","4","5"),
    ordered = TRUE
  )
  x
}

f12 <- function (x){
  x <- factor(
    x,
    levels=c("1","2","3","4"),
    ordered = TRUE
  )
  x
}

f13 <- function (x){
  x <- factor(
    x,
    levels=c("1","2","3","4","5","6"),
    ordered = TRUE
  )
  x
}

```


## Main processing

```{r}
Data2 <- data
data[] <- lapply(data, as.factor)

tmp <- data %>% 
  mutate_at(vars(which(str_detect(colnames(data), "^OpinionBloodService[0-9]$|^IronSuppSanq[0-9]$"))), ~ f1(.)) %>%
  mutate_at(vars(which(str_detect(colnames(data), "^Symptoms([0-9]|10)$"))), ~ f2(.)) %>%
  mutate_at(vars(which(str_detect(colnames(data), "^Measurement[0-9]$|^SupplementsUse.NL$|^SupplementsOpen$"))), ~ f3(.)) %>%
  mutate_at(vars(which(str_detect(colnames(data), "^Knowledge[0-9]$|^KnowledgeHb[0-9]$|^KnowledgeFer[0-9]$|^IronSuppKnowledge[0-9]$"))), ~ f4(.))  %>%
  
  mutate_at(vars(which(str_detect(colnames(data), "_Certainty"))), ~ f5(.)) %>%
  mutate_at(vars(which(str_detect(colnames(data), "^AccForIronLoss$|^AccForIronLoss([0-9]|4A)$|^DeferralSource[0-9](A|_ARC)?$|^SupplementsUse_Fe_Reason[1-9]$"))), ~ f6(.))  %>%
  mutate_at(vars(which(str_detect(colnames(data), "^ConsequenceHb[1-9](_ARC)?$|^ConsequenceFer[1-9](_ARC)?$|^LackDeferralInfo(_lowFer)?[0-9]$|^SymptomsID[0-9]$|^IronSuppUse[1-9]$|^IronSuppSE[1-9]$"))), ~ f7(.)) %>%
  mutate_at(vars(matches("IronSuppEffect[1-6A]")), ~ f8(.))%>%
  mutate_at(vars(matches("IronSuppSanqCondition[1-9]")), ~ f9(.)) %>%
  mutate_at(vars(which(str_detect(colnames(data), "_Confidence.FI"))), ~ f11(.))%>%
  mutate_at(vars(which(str_detect(colnames(data), "^IronDefPref[0-9]$"))), ~ f12(.)) %>%
  mutate_at(vars(which(str_detect(colnames(data), "^IronSuppSanqReason[0-9](_ARC)?$"))), ~ f13(.))

  
tmp <- tmp %>%  mutate( Age = fct_recode(Age,
                                         "18-19"="1",
                                         "18-19"="2",
                                         "20-29"="3",
                                         "20-29"="4",
                                         "20-29"="5",
                                         "20-29"="6",
                                         "20-29"="7",
                                         "20-29"="8",
                                         "20-29"="9",
                                         "30-39"="10",
                                         "30-39"="11",
                                         "40-49"="12",
                                         "40-49"="13",
                                         "50-59"="14",
                                         "50-59"="15",
                                         "60-69"="16",
                                         "60-69"="17",
                                         "70-79"="18",
                                         "70-79"="19",
                                         "80+"="20"),
    Sex =  fct_recode(Sex,  
                  "Male" = "1",
                  "Female" = "2",
                  "Non-binary" = "3",
                  "Other"="4",
                  "Prefer not to say" = "5"
                  ),
    
    Area = fct_recode(Area,
                     "In a city of over 100,000 people" = "1",
                     "In a city of over 100,000 people" = "2",
                    "In a town of less than 100,000 people, or in the countryside" = "3"
                                                 
    ),
        Menstruation = fct_recode(Menstruation,  
    "Yes" = "1",
    "No, due to current pregnancy" = "2",
    "No, due to menopause" = "3",
    "No, due to current contraception use" = "4"
     ),  
      Education =fct_recode(Education,
"None" = "0",
"Primary school" = "1",
"Primary school" = "2", 
"High school" = "3",
"College" = "4",
"College" = "5",
"Applied sciences" = "10",
"University" = "6",
"University" = "7",
"Other" = "8"
      ) %>% fct_relevel(
"Other",
"None",
"Primary school", 
"High school",
"College",
"Applied sciences",
"University"
      ) %>% as.ordered(.)
,
    Health = fct_recode(Health,
"Excellent" =  "1",
"Very good" =  "2",
"Good" =   "3",
"Fair" =  "4",
"Poor" =  "5"
) %>% fct_relevel(.,
"Poor",
"Fair",
"Good",
"Very good",
"Excellent",  
) %>% as.ordered(.)
,
    Diet = as.factor(Diet)%>%
  #str_replace_all(data$Diet,"\\(|\\)","_") %>% 
    fct_recode(.,
      "Yes, vegan (no animal products)" =  "1",
"Yes, vegetarian (no meat and no fish)" =  "2",
"Yes, pescatarian (no meat)" =  "3",
"Yes, flexitarian (at least three days without meat)" = "4" ,
"No" = "5"

    )
,
Deferral
=fct_recode(Deferral,
            "No" = "3",
            "I don’t know"= "4",
            "Yes, once "= "1",
            "Yes, multiple times "= "2"
            
)
,
Deferral_hb
=fct_recode(Deferral_hb,
            "No" = "3",
            "I don’t know"= "4",
            "Yes, once "= "1",
            "Yes, multiple times "= "2"
            
),
Deferral_lowFer
=fct_recode(Deferral_lowFer,
            "No" = "3",
            "I don’t know"= "4",
            "Yes, once "= "1",
            "Yes, multiple times "= "2"
            
),
Number_donation = fct_recode(Number_donation,
      "1 - 3 times" = "1",
      "1 - 3 times" = "2",
      "1 - 3 times" = "3",
      "1 - 3 times" = "4",
"4 - 10 times" = "5",
"4 - 10 times" = "6",
"4 - 10 times" = "7",
"More than 10 times" = "8",
"None" = "13",
"I don't know" = "12",
"More than 20 times" = "9",
"More than 20 times" = "10",
"More than 20 times" = "11"
    ) %>% fct_relevel(.,
    "I don't know",
    "None",
    "1 - 3 times",
"4 - 10 times",
"More than 10 times",
"More than 20 times"
) %>% as.ordered(.)
,
Number_donation_pp_p2 = fct_recode(Number_donation_pp_p2,
      "1 - 3 times" = "1",
      "1 - 3 times" = "2",
      "1 - 3 times" = "3",
      "1 - 3 times" = "4",
"4 - 10 times" = "5",
"4 - 10 times" = "6",
"4 - 10 times" = "7",
"More than 10 times" = "8",
"None" = "13",
"I don't know" = "12",
"More than 20 times" = "9",
"More than 20 times" = "10",
"More than 20 times" = "11"
    ) %>% fct_relevel(.,
    "I don't know",
    "None",
    "1 - 3 times",
"4 - 10 times",
"More than 10 times",
"More than 20 times"
) %>% as.ordered(.)
,
    Last_donation = fct_recode(Last_donation,
"Within the last 6 months" =  "1",
"6 months to a year ago" =  "2",
"1-2 years ago" =  "3",
"More than 2 years ago" =  "4",
"I cannot remember" = "5",
    ) %>% fct_relevel(.,
"Within the last 6 months",
"6 months to a year ago",
"1-2 years ago",
"More than 2 years ago",
"I cannot remember"
                      ) %>% as.ordered(.)
,
Last_Deferral = fct_recode(Last_Deferral,
"Currently" = "0",
"Within the last 6 months" =  "1",
"6 months to a year ago" =  "2",
"1-2 years ago" =  "3",
"More than 2 years ago" = "4",
"I cannot remember" = "5",
    ) %>% fct_relevel(.,
                      "Currently",
"Within the last 6 months",
"6 months to a year ago",
"1-2 years ago",
"More than 2 years ago",
"I cannot remember"
                      ) %>% as.ordered(.)
,
Last_Deferral_hb = fct_recode(Last_Deferral,
"Currently" = "0",
"Within the last 6 months" =  "1",
"6 months to a year ago" =  "2",
"1-2 years ago" =  "3",
"More than 2 years ago" = "4",
"I cannot remember" = "5",
    ) %>% fct_relevel(.,
                      "Currently",
"Within the last 6 months",
"6 months to a year ago",
"1-2 years ago",
"More than 2 years ago",
"I cannot remember"
                      ) %>% as.ordered(.)
,
Last_Deferral_lowFer = fct_recode(Last_Deferral,
"Currently" = "0",
"Within the last 6 months" =  "1",
"6 months to a year ago" =  "2",
"1-2 years ago" =  "3",
"More than 2 years ago" = "4",
"I cannot remember" = "5",
    ) %>% fct_relevel(.,
                      "Currently",
"Within the last 6 months",
"6 months to a year ago",
"1-2 years ago",
"More than 2 years ago",
"I cannot remember"
                      ) %>% as.ordered(.)
,

   
IronSuppSanqReasonTop =
  fct_recode(IronSuppSanqReasonTop,
             "To improve my health" = "2",
             "To be able to donate without delays"    = "1"
  )
  )
  
```


# Deal with open text answers

# remove the open text answers for making the summary
```{r}
#most were already removed

tmp <- tmp %>% select(-IronSuppSanqReason6)
```


# Export summary data

```{r}
#Export summary count data
DataExport <- apply(tmp,MARGIN = 2,FUN = table)
saveRDS(DataExport,file="L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/2. Datasets/individual/1. iDPS - Summary data - USA.RDS")

```


```{r}
# dfSummary(data,
#           plain.ascii = FALSE, 
#           style = "grid", 
#           graph.col = TRUE,
#           graph.magnif = 0.75, 
#           valid.col = FALSE, 
#           tmp.img.dir = "/tmp")


datausa <- saveRDS(tmp, file = "L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/A Source documents/iDPS - USA/datausa.RDS")

```
