==> src/2_summarise_data_FR.Rmd <==
---
title: "Donor iron supplement perception data summary FR"
author: "Amber Meulenbeld"
date: "`r Sys.time()`"
output:
  html_document:
    toc: true
    theme: united
---

```{r include=FALSE}
#run code below in console:
 # rmarkdown::render('L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/1. R Scripts/src/2_summarise_data_FR.Rmd', clean=TRUE,output_format='html_document',output_file='L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/4. Results/summarise_data_FR.html') 

knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	warning = FALSE
)
library(tidyverse)
library(gtsummary)
library(likert)
library(sjlabelled)
library(readxl)
library(ggtext)

```



```{r}
#data <- readRDS("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/2. Datasets/1. iDPS - Summary data - FR.RDS")
data <- readRDS("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/A Source documents/iDPS - FR/dataFR.rds")

#We only want to look at blood donors
# data <- data %>% 
#   filter(Number_donation != "Unknown")

questions <- read_excel("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/A Source documents/iDPS - Variable list/DPS Variable Overview - for summary.xlsx")
questions <- questions %>% filter(FR == "x")
mapping <- setNames(questions$Question, questions$`Variable name`)
```

# Demographics

## Age

```{r}

p <- ggplot(data %>% filter(Sex == "Male" | Sex == "Female"))
p <- p + geom_bar(aes(x = factor(Age)))  # Treat Age as a categorical variable using factor()
p <- p + facet_wrap(. ~ Sex)
p <- p + labs(x = "Age", y = "Count")    # Optionally add labels for clarity
p

```


```{r}
# data2<- data
# colnames(data2) <- str_replace_all(colnames(data2),pattern = " ", "_") %>% 
#   str_replace_all(.,".*:_","") %>% 
#   str_replace_all(.,"\\?","") %>% 
#   str_replace_all(.,",","")

```


```{r}

table2 <-
  tbl_summary(
    data,
    include = c(
      Age,
      Area,
      Education,
      Health,
      Diet
      ),
    #by = What_is_your_gender, # split table by group
    missing = "no" # don't list missing data separately
  ) %>%
  add_n()  # add column with total number of non-missing observations
table2
```


## Donation history

```{r}
table3 <-
  tbl_summary(
    data,
    include = c(
      Number_donation,
      Last_donation,
      ),
    #by = What_is_your_gender, # split table by group
    missing = "no" # don't list missing data separately
  ) %>%
  add_n()  # add column with total number of non-missing observations
table3

```



# Blood service conduct


```{r , warning=FALSE}
items = data %>%
    select(which(str_detect(colnames(data) ,"OpinionBloodService")))


colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

nlevels = 5
```


```{r , warning=FALSE}
q10 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q10, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)
```


# Donation symptoms

## Occurence

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^Symptoms[0-9]$"))) 

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))

# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 4

```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE
     )+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```


## Severity


```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^Symptoms[0-9]_Severity$"))) 
colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))

# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 5

```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

# Donation knowledge

## General



```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^Knowledge[0-9]$"))) %>% mutate_all(~as.factor(.))%>% mutate_all(~ factor(., levels = c("TRUE", "FALSE", "I don't know"), labels = c("TRUE", "FALSE", "I don't know")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

nlevels = 3
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)
```




```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^Knowledge[0-9]_Certainty$"))) %>% mutate_all(~as.factor(.))%>% mutate_all(~ factor(., levels = c("50/50", "100%"), labels = c("50/50", "100%")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\: ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))

# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 2
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```



## Hemoglobin



```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^KnowledgeHb[0-9]$"))) %>% mutate_all(~as.factor(.))%>% mutate_all(~ factor(., levels = c("TRUE", "FALSE", "I don't know"), labels = c("TRUE", "FALSE", "I don't know")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

nlevels = 3
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)
```




```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^KnowledgeHb[0-9]_Certainty$"))) %>% mutate_all(~as.factor(.))%>% mutate_all(~ factor(., levels = c("50/50", "100%"), labels = c("50/50", "100%")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\: ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))

# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 2
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```





## Ferritin



```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^KnowledgeFer[0-9]$"))) %>% mutate_all(~as.factor(.))%>% mutate_all(~ factor(., levels = c("TRUE", "FALSE", "I don't know"), labels = c("TRUE", "FALSE", "I don't know")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

nlevels = 3
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)
```




```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^KnowledgeFer[0-9]_Certainty$"))) %>% mutate_all(~as.factor(.))%>% mutate_all(~ factor(., levels = c("50/50", "100%"), labels = c("50/50", "100%")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\: ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))

# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 2
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```


## Measurements

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^Measurement[0-9]$"))) %>% mutate_all(~as.factor(.))%>% mutate_all(~ factor(., levels = c("Yes", "No", "I don't know"), labels = c("Yes", "No", "I don't know")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 3
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```



## Process


```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^ConsequenceHb[0-9]$"))) %>% mutate_all(~ replace_na(., 0), ~ as.factor(.)) %>% mutate_all(~ factor(., levels = c(FALSE, TRUE), labels = c(FALSE, TRUE)))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 2

```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

# Hb-deferral

## Occurence

```{r}

data %>% count(Deferral)

p <- ggplot(data %>% filter( Sex== "Male" | Sex == "Female")) +  geom_histogram(aes(x=Deferral), stat = 'count')
p
```

```{r}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^Deferral$"))) %>% mutate_all(~ as.factor(.))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

nlevels = 4
```

```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)
```

## Last time

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"Last_Deferral"))) %>% mutate_all(~ as.factor(.))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

nlevels = 6

```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)
```


## Was the information about deferral ...

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^DeferralInfo[0-9]$")))  %>% mutate_all(~ as.factor(.))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\: ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers


nlevels = 5
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

## Missing information

```{r}
cols <- c("LackDeferralInfo1", "LackDeferralInfo2", "LackDeferralInfo3","LackDeferralInfo4", "LackDeferralInfo5")
cols_include <- cols %in% colnames(data)
table2 <-
  tbl_summary(
    data,
    include = cols[cols_include],
    #by = What_is_your_gender, # split table by group
    missing = "no" # don't list missing data separately
  ) %>%
  add_n()  # add column with total number of non-missing observations
table2


```

```{r}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^LackDeferralInfo[0-9]$")))  %>% mutate_all(~ replace_na(., 0), ~as.factor(.)) %>% mutate_all(~ factor(., levels = c(FALSE, TRUE), labels = c(FALSE, TRUE)))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 2
```

```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

## Effect to motivation

1= no effect, 5 = much effect

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^Deferral_Motivation[0-9]$"))) 

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 5
```

# 1 = no effect, 5 = much effect

```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```
## Hypothetical effect on motivation

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^Deferral_Motivation_Hypothetical[0-9]$"))) 

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 5
```

# 1 = no effect, 5 = much effect

```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

# Iron deficiency 

## Symptoms

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^SymptomsID[0-8]$"))) %>% mutate_all(~ replace_na(., 0), ~as.factor(.)) %>% mutate_all(~ factor(., levels = c(FALSE, TRUE), labels = c(FALSE, TRUE)))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 2
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

## Replacement

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^AccForIronLoss[0-8]$"))) %>% mutate_all(~ replace_na(., 2), ~as.factor(.)) %>% mutate_all(~ factor(., levels = c(FALSE, TRUE), labels = c(FALSE, TRUE)))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 2
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```


## Information

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^DeferralSource[0-9]$"))) %>% mutate_all(~factor(., levels = c(FALSE, TRUE), labels = c(FALSE, TRUE)))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 2
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```
## Confidence in information source

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^DeferralSource[0-9]_Confidence.FI$"))) 

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 5
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

# Opinion on supplements

## Iron
not inlcuded in FR data
```{r , eval=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^Supplements[0-9]$"))) 

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 5
```


```{r , eval=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

## Dietary


```{r}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^SupplementsUse$"))) 

table2 <-
  tbl_summary(
    items,
    include = c(
      SupplementsUse
            ),
    #by = What_is_your_gender, # split table by group
    missing = "no" # don't list missing data separately
)%>%
  add_n()  # add column with total number of non-missing observations
table2

```




```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^SupplementsReason[0-9]$"))) %>% mutate_all(~factor(., levels = c(FALSE, TRUE), labels = c(FALSE, TRUE)))


colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 2
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

Are there other reasons 
```{r}
data %>% count(SupplementsReason6A)
```

## Iron statements

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^IronSuppKnowledge[0-9]$"))) %>% mutate_all(~factor(., levels = c(FALSE, TRUE), labels = c(FALSE, TRUE)))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 2
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^IronSuppKnowledge[0-9]_Certainty$"))) %>% mutate_all(~as.factor(.))%>% mutate_all(~ factor(., levels = c("50/50", "100%"), labels = c("50/50", "100%")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])

# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 2
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```


## Iron side effects


```{r , warning=FALSE}
items = data%>% 
    select(which(str_detect(colnames(data) ,"^IronSuppSE[0-9]$")))  %>% mutate_all(~ replace_na(., 2), ~as.factor(.)) %>% mutate_all(~ factor(., levels = c(FALSE, TRUE), labels = c(FALSE, TRUE)))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 2

```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

## Iron use

```{r}
items = data%>% 
    select(which(str_detect(colnames(data) ,"^IronSuppUse$")))  %>% mutate_all(~as.factor(.))
table2 <-
  tbl_summary(
    items,
    include = c(
      IronSuppUse
            ),
  #  by = What_is_your_gender, # split table by group
    missing = "no" # don't list missing data separately
  ) %>%
  add_n()  # add column with total number of non-missing observations
table2


```



```{r , warning=FALSE}
items = data %>%
    select(which(str_detect(colnames(data), "^IronSuppEffect[0-9]$"))) %>%
    mutate_all(~ factor(., levels = c("Increased", "No effect", "Decreased"), labels = c("Increased", "No effect", "Decreased")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 3
```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

## Reaction to offering iron
Not included in FR data
```{r , eval=FALSE}
items = data%>% 
    select(which(str_detect(colnames(data) ,"^IronSuppSanq[0-9]$"))) %>% mutate_all(~as.factor(.)) %>% mutate_all(~ factor(., levels = c(1,2,3,4,5), labels = c("Totally disagree","Disagree","Neutral","Agree","Totally agree")))

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\, ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 5
```


```{r , eval=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```



```{r}
items = data%>% 
    select(which(str_detect(colnames(data) ,"^IronSuppSanqReason$"))) %>% mutate_all(~as.factor(.)) %>% mutate_all(~ factor(., levels = c(1,2,3,4), labels = c("To improve my health","To be able to continue to donate","Different reason","I would never make use of iron supplements provided by the blood service")))


items %>% count(IronSuppSanqReason)

```

## Reason to use iron

```{r , warning=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^IronDefPref[0-9]$"))) 

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers

nlevels = 4

```


```{r , warning=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

"Please drag the options in the right order, with 1 being most preferred and 4 least preferred option"



## Willingness to use iron
Not included in FR data
```{r , eval=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^IronSuppSanqCondition[0-9]$"))) %>% mutate_all(~as.factor(.)) %>% mutate_all(~ factor(., levels = c("Disagree", "Agree"), labels = c("Disagree", "Agree")))


colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 2

```


```{r , eval=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

## Bases for policy

Not included in FR data
```{r , eval=FALSE}
items = data %>% 
    select(which(str_detect(colnames(data) ,"^IronSuppSanqReason[0-9]$"))) 

colnames(items) <- unname(mapping[match(colnames(items), names(mapping))])
# Separate the question and answer
questions_and_answers <- colnames(items)

# Split the column names to isolate the answers
split_colnames <- strsplit(questions_and_answers, "\\? ") # Split by the question mark followed by a space

# Extract the common question (assuming all columns share the same question)
question <- unique(sapply(split_colnames, `[`, 1))
question <- question[1]
# Extract only the answers (the part after the question)
answers <- sapply(split_colnames, `[`, 2)

# Update the column names with just the answers
colnames(items) <- answers
nlevels = 5
```


```{r , eval=FALSE}
q11 <- likert(
  items = as.data.frame(items),
  nlevels = nlevels
  
)
plot(q11, 
     type = "heat",
     legend.position = "bottom",
     text.size=3,
     include.histogram = FALSE)+ ggtitle(question)+
  theme(plot.title = element_textbox_simple())
```

"Please drag the options in the right order, with 1 being most preferred and 5 least preferred option"
