==> src/5_summarise_data_across_countries.Rmd <==
---
title: "Donor iron supplement perception data summary across countries"
author: "Amber Meulenbeld"
date: "`r Sys.time()`"
output:
  html_document:
    toc: true
    theme: united
---

```{r setup, include=FALSE}
#rmarkdown::render("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/1. R Scripts/src/5_summarise_data_across_countries.Rmd", clean=TRUE,output_format='html_document',output_file='L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/4. Results/summarise_data_across_countries.html')
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	warning = FALSE
)
library(tidyverse)
library(readxl)
library(dplyr)

```

```{r functions, include=FALSE}
# Helper function to save plots
save_plot <- function(p, filename, width, height) {
  ggsave(filename, plot = p, width = width, height = height)
}

plot_type_1 <- function(qdata=qdata, country_colors, folder_name, VariableName, figwidth=10, figheight=3, free_y=T) {

 p <- ggplot(qdata) + 
  geom_col(aes(x = answer, y = Percentage, fill = Country)) +  # Reverse x and y to correct orientation
  coord_flip()  # Flip the axes to maintain the correct orientation
if(free_y){
  p <- p + facet_wrap(~ Country, scales = "free_y", strip.position = "top", nrow = 1)
} else {
  p <- p + facet_wrap(~ Country, strip.position = "top", nrow = 1)
}
p <- p + 
  xlab("Percentage") +
  ylab(" ") +
  ggtitle(unique(qdata$Question_Text)[1]) +
  scale_y_continuous(breaks = seq(0, 100, by = 50), limits = c(0, 100)) +  # Note the y-axis is now continuous
  scale_fill_manual(values = country_colors) +
  guides(fill = FALSE) +
  theme_bw() +
  theme(
    strip.text.y.left = element_text(angle = 0),
    strip.placement = "outside",
    strip.text = element_text(size = 12)
  ) +
  geom_text(aes(x = answer, y = Percentage, label = round(Percentage, 1)), 
            hjust = 0, size = 3) +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 10))  # x-axis is now the discrete axis

print(p)
#   # Create plot
#   p <- ggplot(qdata) + 
#     geom_col(aes(x = Percentage, y = answer, fill = Country)) 
#   if(free_y){
#     p <- p+facet_wrap(~ Country, scales = "free_y", strip.position = "top", nrow = 1)
#   }else{
#     p <- p+facet_wrap(~ Country, strip.position = "top", nrow = 1)
#   }
#     p<- p + xlab("Percentage") +
#     ylab(" ") +
#     ggtitle(unique(qdata$Question_Text)[1]) +
#     scale_x_continuous(breaks = seq(0, 100, by = 50), limits = c(0, 120)) +
#     scale_fill_manual(values = country_colors) +
#     guides(fill = FALSE) +
#     theme_bw() +
#     theme(
#       strip.text.y.left = element_text(angle = 0),
#       strip.placement = "outside",
#       strip.text = element_text(size = 12)
#     ) +
#     geom_text(aes(x = Percentage, y = answer, label = round(Percentage, 1)), 
#               hjust = -0.5, size = 3) +
#     scale_y_discrete(labels = function(x) str_wrap(x, width = 10))
#   
#   print(p)
# 
   # Save plot
   filename <- paste0(folder_name, "/", VariableName, " - by country.png")
   save_plot(p, filename, width = figwidth, height = figheight)
}


plot_type_2 <- function(qdata=qdata, country_colors=country_colors, folder_name=FolderName, VariableName, figwidth=9, figheight=7) {

  # Create plot
  p <- ggplot(qdata) +
    geom_bar(aes(x = Statement, y = Percentage, fill = Country), position = 'dodge', stat = "identity") +
    ylab("Percentage") +
    xlab(" ") +
    ggtitle(str_wrap(paste0(unique(qdata$Q)[1], "?"), width = 75)) +
    scale_y_continuous(breaks = seq(0, 100, by = 50), limits = c(0, 101), position = "left") +
    scale_fill_manual(values = country_colors) +
    theme_bw() +
    theme(
      strip.text.y.left = element_text(angle = 0),
      strip.placement = "outside"
    ) +
    geom_text(aes(x = Statement, y = Percentage, label = round(Percentage, 1), group = Country), 
              vjust = -0.5, size = 3, 
              position = position_dodge(width = 0.9)) +
    scale_x_discrete(labels = function(x) str_wrap(x, width = 10))
  
  print(p)

  # Save plot
  filename <- paste0(folder_name, "/", VariableName," - by country.png")
  save_plot(p, filename, width = figwidth, height = figheight)
}

plot_type_3 <- function(qdata=qdata, country_colors=country_colors, folder_name=FolderName, VariableName, free_x=T, figwidth=9, figheight=7) {
  #this function is for plots with the percentage on the x axis and the categories on the x axis, faceted by country and statement
  p <- ggplot(qdata) + 
    geom_col(aes(x = answer, y = Percentage, fill = Country)) 
  if(free_x){
    p <- p+    facet_grid(Statement ~ Country, 
               labeller = labeller(Statement = label_wrap_gen(width = 25, multi_line = TRUE)),
               switch = "y",
               scales = "free_x") 
  }else{
        p <- p+    facet_grid(Statement ~ Country, 
               labeller = labeller(Statement = label_wrap_gen(width = 25, multi_line = TRUE)),
               switch = "y") 
  }

    p <- p + ylab("Percentage") +
    xlab(" ") +
    ggtitle(str_wrap(unique(qdata$Q)[1], width = 75)) +
    scale_y_continuous(breaks = seq(0, 100, by = 50), limits = c(0, 120), position = "right") +
    scale_fill_manual(values = country_colors) +
    guides(fill = FALSE) +
    theme_bw() +
    theme(
      strip.text.y.left = element_text(angle = 0),
      strip.placement = "outside"
    ) +
    geom_text(aes(x = answer, y = Percentage, label = round(Percentage, 1)), 
              vjust = -0.5, size = 3) +
    scale_x_discrete(labels = function(x) str_wrap(x, width = 10))
  
  print(p)
  
  # Save plot
  filename <- paste0(folder_name, "/", VariableName, " - by country.png")
  save_plot(p, filename, width = figwidth, height = figheight)
}

plot_type_3a <- function(qdata=qdata, folder_name=FolderName, VariableName, free_x=T, figwidth=9, figheight=7, legendtitle=legendtitle) {
  p <- ggplot(qdata) + 
    theme_bw() +
    geom_col(aes(x = answer, y = Percentage, fill = interaction(Country, VariableGroup)), position = position_dodge(width = 0.9))  
  
  if(free_x){
    p <- p + facet_grid(Statement ~ Country, 
                        labeller = labeller(Statement = label_wrap_gen(width = 25, multi_line = TRUE)),
                        switch = "y", scales = "free_x") 
  } else {
    p <- p + facet_grid(Statement ~ Country, 
                        labeller = labeller(Statement = label_wrap_gen(width = 25, multi_line = TRUE)),
                        switch = "y") 
  }

  p <- p + ylab("Percentage") +
    xlab(" ") +
    ggtitle(str_wrap(unique(qdata$Q)[1], width = 75)) +
    scale_y_continuous(breaks = seq(0, 100, by = 50), limits = c(0, 120), position = "right") +
    scale_fill_manual(name = legendtitle, breaks = c("USA.hb", "USA.fer"), labels = c("Hb", "Ferritin"), values = c(
      "Netherlands.other" = "#E69F00", 
      "France.other"      = "#0072B2", 
      "Sweden.other"      = "#F0E442", 
      "Finland.other"     = "#56B4E9", 
      "Japan.other"       = "#D55E00",  
      "Germany.other"     = "#CC79A7", 
      "USA.hb" = "#B3B300",
      "USA.fer" = "#808000"
    )) +
    # Enable the legend to check grouping
    # guides(fill = FALSE) +
    theme(strip.text.y.left = element_text(angle = 0),
          strip.placement = "outside") +
    geom_text(aes(x = answer, y = Percentage, label = round(Percentage, 1), 
                  group = interaction(Country, VariableGroup)), 
              vjust = -1.5, size = 2, color = "black", position = position_dodge(width = 0.9)) +
    scale_x_discrete(labels = function(x) str_wrap(x, width = 10))
  
  print(p)

  # Save plot with increased size
  filename <- paste0(folder_name, "/", VariableName, " - by country.png")
  save_plot(p, filename, width = figwidth + 2, height = figheight + 2)
}


plot_type_4 <- function(qdata=qdata, country_colors=country_colors, folder_name=FolderName, VariableName, free_y=T,figwidth=9, figheight=7, Qtitle=T){
  
  p <- ggplot(qdata) + 
  geom_col(aes(x = Percentage, y = answer, fill = Country))   # Swap x and y
  if(free_y){
    p<- p+ facet_grid(Statement ~ Country, 
             labeller = labeller(Statement = label_wrap_gen(width = 30, multi_line = TRUE)),
             switch = "y",
             scales="free_y") 
    } else {
          p<- p+ facet_grid(Statement ~ Country, 
             labeller = labeller(Statement = label_wrap_gen(width = 30, multi_line = TRUE)),
             switch = "y") 
    }# Keeps facet labels on the left
  p <- p + xlab("Percentage") +  # Now label x-axis as "Percentage"
  ylab(" ")   # Y-axis for "answer"
  if(Qtitle){
    p <- p+ggtitle(paste0(unique(qdata$Q)[1], "?")) 
  }else{
    p <- p+ggtitle(unique(qdata$Q)[1])
  }
    
  p <- p + scale_x_continuous(breaks = seq(0, 100, by = 50), limits = c(0, 110), position = "bottom") +  # Ensure x-axis (Percentage) is at the bottom
  scale_fill_manual(values = country_colors) +
  guides(fill = FALSE) +
  theme_bw() +
  theme(
    strip.text.y.left = element_text(angle = 0),  # Keeps facet text horizontal
    strip.placement = "outside",  # Places facet labels outside the plot area
    axis.text.y = element_text(hjust = 1),  # Moves y-axis (answer) labels to the right
    axis.title.y = element_text(hjust = 1)  # Adjusts title of y-axis (answer)
  ) +
  geom_text(aes(x = Percentage, y = answer, label = round(Percentage, 1)), 
            hjust = -0.5, size = 3) +  # Adjust text alignment based on new x and y
  scale_y_discrete(labels = function(x) str_wrap(x, width = 15), position="right")  # Adjust y-axis labels (answer)

print(p)

  filename <- paste0(folder_name, "/", VariableName, " - by country.png")
  save_plot(p, filename, width = figwidth, height = figheight)
} 


```

```{r load data, echo=F}
currentDate <- format(Sys.time(), "%Y-%m-%d")
FolderName <- paste("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/3. Plots/", currentDate,sep="")
dir.create(FolderName)

# Read in combined country data 
data <- readRDS("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/B Analysis/R/2. Datasets/combined.RDS") %>% mutate(country = ifelse(country == "FIN2", "FIN", country)) %>% group_by(Question, country, Response) %>%
    summarise(Count = sum(Count), .groups = 'drop') %>%
    group_by(Question, country) %>%
    mutate(Percentage = 100 * Count / sum(Count)) %>%
    ungroup()

#Change country codes to full names
country_mapping <- c("NL" = "Netherlands", "FR" = "France", "SWE" = "Sweden", 
                     "FIN" = "Finland", "JPN" = "Japan", "USA" = "USA")
data$Country <- country_mapping[data$country]
country_colors <- c(
  "Netherlands" = "#E69F00", # Orange (Colorblind-friendly orange)
  "France"      = "#0072B2", # Blue (Colorblind-friendly blue)
  "Sweden"      = "#F0E442", # Yellow (Colorblind-friendly yellow)
  "Finland"     = "#56B4E9", # Sky blue (Colorblind-friendly cyan)
  "Japan"       = "#D55E00",  # Red (Colorblind-friendly reddish-brown)
  "Germany"     = "#CC79A7", # Purple (Colorblind-friendly purple)
  "USA" = "#B3B300" # Olive (Colorblind-friendly olive?)
)


#read in questions
question_text <- read.csv("L:/DonorMedicin/FORTE/E Internation Donor Perception Survey/A Source documents/iDPS - Variable list/questions.csv") %>% select(c(Variable, Question_Text)) %>% distinct(Variable, Question_Text, .keep_all = TRUE)


```

# Characteristics

## Age


```{r, echo=F, fig.width = 10, fig.height = 3}

q <- data %>% filter(grepl("^Age$", Question))

#manually add row for NL for aesthetic reasons
new_row <- data.frame(Question = "Age", country = "NL", Response = "18-20", 
                      Count = 0, Percentage = 0, Country = "Netherlands", stringsAsFactors = FALSE)

q <- rbind(q, new_row)


qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%  filter(Response %in% c("18-20", "17-20", "21-30", 
                                          "31-40", "41-50", 
                                          "51-60", "61-70", "71-80","18-19","20-29","30-39","40-49","50-59","60-69","70-79", "80+")) %>%
  mutate(
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(Response, levels = c("18-20", "17-20", "21-30", 
                                          "31-40", "41-50", 
                                          "51-60", "61-70", "71-80","18-19","20-29","30-39","40-49","50-59","60-69","70-79", "80+"))
  )

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Age", folder_name = FolderName)
```

## Sex

```{r, echo=F,fig.width = 10, fig.height = 3}
q <- data %>% filter(grepl("^Sex$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable") %>%  
  mutate(
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(Response, levels = c("Prefer not to say", "Other", "Non-binary", "Male", "Female" ))
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Sex", folder_name = FolderName, free_y=F)
```

## Menstruation

```{r, echo=F,fig.width = 10, fig.height = 5}
q <- data %>% filter(grepl("^Menstruation$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = ifelse(Response == "No, other", "No, due to a different reason", as.character(Response)),
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(Response, levels = c("Yes",
                                         "No, due to current contraception use",
                                         "No, due to current pregnancy", 
                                         "No, due to hysterectomy",
                                         "No, due to menopause",
                                         "No, due to a different reason",
                                         "No"))
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Menstruation", folder_name = FolderName, free_y=F, figheight = 5)
```

## Area

```{r, echo=F,fig.width = 10, fig.height = 3}
q <- data %>% filter(grepl("^Area$", Question))



qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = ifelse(Response=="In a city of over 100 000 people", "In a city of over 100,000 people", as.character(Response)),
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = Response
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Area", folder_name = FolderName, free_y=F)
```

## Education

```{r, echo=F,fig.width = 10, fig.height = 5}
q <- data %>% filter(grepl("^Education$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = ifelse(Response == "Applied Sciences", "Applied sciences", as.character(Response)),
    Response = ifelse(Response == "High School", "High school", as.character(Response)),
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(Response, levels = c("None", "Other","Primary school", "High school", "College", "Applied sciences", "University"))
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Education", folder_name = FolderName, free_y=F, figheight = 5)
```

## Health

```{r, echo=F,fig.width = 10, fig.height = 5}
q <- data %>% filter(grepl("^Health$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(Response, levels = c("Poor", "Fair", "Good", "Very good", "Excellent"))
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "HealthStatus", folder_name = FolderName, free_y=F, figheight = 5)
```

## Diet
```{r, echo=F,fig.width = 10, fig.height = 5}
q <- data %>% filter(grepl("^Diet$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(Response, levels = c(
      "No",                                                          
"Yes, flexitarian (at least three days without meat)",        
"Yes, lakto-ovovegan (no meat and no fish, but milk and eggs)",
"Yes, pescatarian (no meat)",                                  
"Yes, vegan (no animal products)",                             
"Yes, vegetarian (no meat and no fish)"
    ))
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Diet", folder_name = FolderName, free_y=F, figheight = 5)
```

# Donation behavior

## Number of donations

```{r, echo=F,fig.width = 10, fig.height = 5}
q <- data %>% filter(grepl("^Number_donation$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(Response, levels = c(
"None",
"1 - 3 times",        
"4 - 10 times",       
"More than 10 times", 
"More than 20 times",
"More than 30 times",
"I don't know"
    ))
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Number_donation", folder_name = FolderName, free_y=F, figheight = 5)
```

## Last donation

```{r, echo=F,fig.width = 10, fig.height = 5}
q <- data %>% filter(grepl("^Last_donation$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = ifelse(Response == "More than 2 years ago", "Longer than 2 years ago", as.character(Response)),
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(Response, levels = c(
"I cannot remember",        
"Longer than 2 years ago",  
"1-2 years ago",
"6 months to a year ago",
"Within the last 6 months"
    ))
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Last_donation", folder_name = FolderName, free_y=F, figheight = 5)
```

# Blood service

## Opinion

```{r, echo=F,fig.width = 9, fig.height = 10}
q <- data %>% filter(grepl("^OpinionBloodService([1-9])$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    
    Q = "To what extent do you agree or disagree with the following statements?",
    answer = ordered(Response, levels=c("Totally disagree", "Disagree", "Neutral", "Agree", "Totally agree")),
    Statement=Question_Text
  )%>% 
  droplevels()


plot_type_4(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "OpinionBloodService", free_y=F, figheight = 10)

```

# Deferral

## Been deferred?

```{r, echo=F,fig.width = 10, fig.height = 3}
q <- data %>% filter(grepl("^Deferral$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(trimws(Response), levels = c("I don’t know", "No", "Yes, multiple times", "Yes, once", "Yes"))
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Deferral", folder_name = FolderName, free_y=F)
```

## Last deferral

```{r, echo=F,fig.width = 10, fig.height = 5}
q <- data %>% filter(grepl("^Last_Deferral$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
   mutate(Response = case_when(
    Response %in% c("<6 months", "Within the last 6 months", "Currently") ~ "<6 months",
    Response %in% c("6-12 months", "6 months to a year ago") ~ "6-12 months",
    Response %in% c("12-24 months", "1-2 years ago") ~ "12-24 months",
    Response %in% c(">24 months", "More than 2 years ago") ~ ">24 months",
    Response == "I cannot remember" ~ "I don't know",
    Response == "I don't know" ~ "I don't know",
    TRUE ~ "Other"
  ))%>% mutate(
    #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = factor(Response, levels = c(
"I don't know",
">24 months", 
"12-24 months",            
"6-12 months",   
"<6 months"
    ))
  )%>% 
  droplevels()

plot_type_1(qdata=qdata, country_colors = country_colors, VariableName = "Last_deferral", folder_name = FolderName, free_y=F, figheight = 5)
```

## Deferral source

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("DeferralSource[1-7](A|.NL|_ARC)?$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("TRUE", "FALSE", "Yes", "No"))
  ) %>% filter(Response == "Yes" | Response == "TRUE") %>% mutate(Statement = ordered(Statement, levels = c(" Blood service staff", " Blood service website and/or brochure", " Donor physician"," Health care practitioner", " Search information on the internet", " Friend or family member", " Other, namely", " None")))%>% 
  droplevels()

plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "DeferralSource")
```

## Deferral source confidence

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^DeferralSource[1-9](_ARC)?_Confidence.FI$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = trimws(as.character(Response)),
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Statement = sapply(Statement, function(x) str_remove(x, "confidence |Confidence ")),
    Q = "To what extent do you trust this source of information? (1 = very little, 5 = very much)", #sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c(1,2,3,4,5)) 
  )%>% mutate(Statement = ordered(Statement, levels = c(" Blood service staff", " Blood service website and/or brochure", " Donor physician"," Health care practitioner", " Search information on the internet", " Friend or family member", " Other, namely"))) %>%
  droplevels()

plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="DeferralSource_Confidence", free_x=T)
```


## Deferral Info

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^DeferralInfo(_hb|_lowFer)?[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c(1,2,3,4,5))
  )%>% mutate(VariableGroup = case_when(
    grepl("^DeferralInfo_hb", Question) ~ "hb",
    grepl("^DeferralInfo_lowFer", Question) ~ "fer",
    TRUE ~ "other"
  )) %>% arrange(country)%>%
  droplevels()

plot_type_3a(qdata=qdata, folder_name=FolderName, VariableName="DeferralInfo", free_x=F,legendtitle="Deferral reason")


```

## Missing info about deferral

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^LackDeferralInfo(_hb[1-5]|_lowFer[1-5]|[1-5])$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("TRUE", "FALSE", "Yes", "No", "1", "0"))
  ) %>% filter(Response == "Yes" | Response == "TRUE" | Response =="1") %>% 
  droplevels()


plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "LackDeferralInfo")
```

# Supplements

## General supplements


```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^Supplements[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c(1,2,3,4,5))
  )%>% 
  droplevels()


plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="Supplements", free_x=F)


```

## General use

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^(SupplementsUse.NL|SupplementsUse_Other.FI)$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = Response,
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("Yes", "No", "I don't know", "Never", "Not anymore", "Sometimes", "Every day"))
  ) %>% 
  droplevels()


plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "SupplementsUse_general")
```

## Reason supplement use

Plot not available yet due to data processing issues.

```{r, echo=F,fig.width = 9, fig.height = 7} 
# have to look at data structure again, something is not right at least in NL

# q <- data %>% filter(grepl("^SupplementsReason([0-9]|6A)$", Question))
# 
# qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
#   mutate(
#     Statement = Response,
#     Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
#     answer = ordered(Response, levels=c("TRUE", "FALSE", "Yes", "No"))
#   ) %>% filter(Response == "Yes" | Response == "TRUE") %>% 
#   droplevels()
# 
# 
# plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "SupplementsReason")
```

## Supplements Open

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^SupplementsOpen$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = Response,
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("Yes", "No"))
  ) %>% 
  droplevels()


plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "SupplementsOpen")
```

# Iron Loss

## Account for Iron Loss

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^AccForIronLoss$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = Response,
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("TRUE", "FALSE", "Yes", "No"))
  ) %>% 
  droplevels()


plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "AccForIronLoss")
```

## Account for iron loss: How?

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^AccForIronLoss([1-5]|4A)(\\.FI)?$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("TRUE", "FALSE", "Yes", "No"))
  ) %>% filter(Response == "Yes" | Response == "TRUE") %>% 
  droplevels()


plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "AccForIronLossHow")
```

## Iron supplement use

```{r, echo=F,fig.width = 9, fig.height = 7} 

q <- data %>% filter(grepl("^IronSuppUse$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
   mutate(
     Response = trimws(tolower(Response)),
     Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
     answer=ifelse(str_detect(Response, "multivitamin.*don't know|multivitamin.*do not know"), 
                            "I am taking multivitamins but don't know if they contain iron",
                     ifelse(str_detect(Response, "taken iron supplements.*but not anymore"), 
                            "I have taken iron supplements/multivitamins before, but not anymore",
                     ifelse(str_detect(Response, "never taken iron"), 
                            "No, I have never taken iron (containing) supplements",
                     ifelse(str_detect(Response, "docto"), 
                            "Yes, I am currently taking iron supplements ordered by a doctor",
                     ifelse(str_detect(Response, "currently taking iron supplements.*blood service"), 
                            "Yes, I am currently taking iron supplements provided by the blood service",
                     ifelse(str_detect(Response, "currently taking iron supplements.*drugstore"), 
                            "Yes, I am currently taking iron supplements from a drugstore",
                     ifelse(str_detect(Response, "multivitamins.*contain iron"), 
                            "Yes, I take multivitamins that contain iron",
                     Response))))))),
     Statement=answer
   ) %>% 
   droplevels()
 
 
plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "IronSuppUse")
```

## Concerns Iron Supplements

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^ConcernIronSupp([1-9]|5A)(\\.FI)?$", Question)) %>% mutate(Question = ifelse(country=="FIN", str_remove(Question, ".FI"), Question), Question = ifelse(Question == "ConcernIronSupp5","ConcernIronSupp5A",Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("TRUE", "FALSE", "Yes", "No"))
  ) %>% filter(Response == "TRUE" | Response == "Yes") %>% 
  droplevels()


plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "ConcernsIronSupplements")
```

## Important aspects of iron supplementation policy

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^IronSuppSanqReason[1-9](\\.FI)?$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = trimws(Response),
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Statement = fct_rev(Statement),
    Q = sapply(sapply(str_split(Question_Text, "[:?]"), function(x) x[1]), function(x) paste0(x, "?")),
    answer = ordered(Response, levels=c(1,2,3,4,5,6))
  )%>% 
  droplevels()


plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="ImportantAspectsIronSupp", free_x=T, )

```

## Reason Supplements

Plot not available yet due to data processing issues.

```{r, echo=F}
#Still too much work on this variable homogenization to print it now
# q <- data %>% filter(grepl("^IronSuppSanqReason$", Question)
# 
# 
# qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
#   mutate(
#     #Statement = sub(".*?, then (.*)", "\\1", Question_Text),
#     #Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
#     answer = factor(Response, levels = c("Prefer not to say", "Other", "Non-binary", "Male", "Female" ))
#   )%>% 
#  droplevels()
# 
# #Faceted by country
# p <- ggplot(qdata) + 
#   geom_col(aes(x = Percentage, y = answer, fill = Country)) +  # Swapped x and y
#   facet_wrap(~ Country, strip.position = "top", nrow=1) +  # Use facet_wrap and free_y
#   xlab("Percentage") +  # Label for x-axis (Percentage)
#   ylab(" ") +           # Empty y label
#   ggtitle(unique(qdata$Question_Text)[1]) +
#   scale_x_continuous(breaks = seq(0, 100, by = 50), limits = c(0, 120)) +  # Set x scale for Percentage
#   scale_fill_manual(values = country_colors) +
#   guides(fill = FALSE) +
#   theme_bw() +
#   theme(
#     strip.text.y.left = element_text(angle = 0),  # Keeps facet text horizontal
#     strip.placement = "outside",                  # Places facet labels outside
#     strip.text = element_text(size = 12)          # Adjust facet label size for clarity
#   ) +
#   geom_text(aes(x = Percentage, y = answer, label = round(Percentage, 1)), 
#             hjust = -0.5, size = 3) +  # Adjust for horizontal text justification
#   scale_y_discrete(labels = function(x) str_wrap(x, width = 10))  # Set y scale with wrapped text
# 
# p
# 
# FileName <- paste(FolderName, "/Reason supplements - by country.png",sep="")
# ggsave(FileName, width = 10, height = 3)
```

## Iron supplement effects

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^IronSuppEffect([1-9]|6A)$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = trimws(Response),
    Response = ifelse(Response == "Increases", "Increased", as.character(Response)),
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Statement = ordered(Statement, levels = c(" Physical fitness"," Fatigue"," Dizziness"," Headaches"," Gastrointestinal problems", " Other, namely")),
    Q = "In your experience taking iron supplements, have noticed any effects in relation to the listed conditions?",#sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("Increased", "No effect","Decreased", "Don't know"))
  )%>% 
  droplevels()


plot_type_4(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="IronSuppEffect", free_y=T, figheight=10, Qtitle=T)

```

## Perception iron supplements as a policy

```{r, echo=F,fig.width = 9, fig.height = 10}
q <- data %>% filter(grepl("^IronSuppSanq([1-9])$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = trimws(Response),
    Statement = sapply(sapply(str_split(Question_Text, "[,:?]"), function(x) x[2]), function(x) str_remove(x, "then ")),
    Q = sapply(sapply(str_split(Question_Text, "[,:?]"), function(x) x[1]), function(x) paste0(x, " then...")),
    answer = ordered(Response, levels=c("Totally disagree", "Disagree", "Neutral", "Agree", "Totally agree"))
  )%>% 
  droplevels()


plot_type_4(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "PerceptionIronSuppPolicy", free_y=F, figheight = 10, Qtitle=F)

```

## Iron supplement conditions

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^IronSuppSanqCondition[0-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = "If the blood service would provide iron supplements to you, in what situation would you be willing to use them",#sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("Agree", "Disagree"))
  ) %>% filter(answer == "Agree")%>%
  droplevels()


plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "IronSupplementConditions")
```

## Ferritin deferral preferences

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^IronDefPref[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = trimws(Response),
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c(1,2,3,4,5))
  )%>% 
  droplevels()


plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="IronDefPref", free_x=T)

```

# Symptoms

## Occurrence

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^Symptoms([1-9]|10)$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("Never", "Sometimes", "Often", "Always" ))
  )%>% 
  droplevels()


plot_type_4(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "SymptomsOccurrence", free_y=F)

```

## Severity

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^Symptoms([1-9]|[1-9][0-9])_Severity$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c(1,2,3,4,5))
  )%>% 
  droplevels()

plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="SymptomsSeverity", free_x=F)

```


# Knowledge

## Blood service

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^Knowledge[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = ordered(Response, levels=c("TRUE", "FALSE", "I don't know"))
  )%>% 
  droplevels()

plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="Knowledge", free_x=T)
```

## Measurement

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^Measurement[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("Yes", "No", "I don't know"))
  )%>% 
  droplevels()


plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="Measurement", free_x=T)
```

## Knowledge Hb

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("KnowledgeHb[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = ordered(Response, levels=c("TRUE", "FALSE", "I don't know"))
  )%>% 
  droplevels()

plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="KnowledgeHb", free_x=T)
```

## Knowledge Ferritin

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("KnowledgeFer[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sub(".*?, then (.*)", "\\1", Question_Text),
    Q = "Are the following statements True or False?", #sub("(.*), then.*", "\\1", Question_Text),
    answer = ordered(Response, levels=c("TRUE", "FALSE", "I don't know"))
  )%>% 
  droplevels()


plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="KnowledgeFer", free_x=T)
```

## Consequence Hb

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^ConsequenceHb[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("TRUE", "FALSE"))
  )%>% 
  droplevels()

plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="ConsequenceHb", free_x=T)
```

## Consequence Fer

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^ConsequenceFer[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = as.character(Response),
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("TRUE", "FALSE")) 
  )%>% 
  filter(as.character(Response) == "TRUE" | as.character(answer) == "FALSE")%>% 
  droplevels()

plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="ConsequenceFer", free_x=T)
```

## Iron supplements

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^IronSuppKnowledge[1-9]$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Response = as.character(Response),
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("TRUE", "FALSE")) 
  )%>% 
  filter(as.character(Response) == "TRUE" | as.character(answer) == "FALSE")%>% 
  droplevels()

plot_type_3(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName="Knowledge Iron Supplements", free_x=T)
```

## Iron Supplement Side Effects

```{r, echo=F,fig.width = 9, fig.height = 7}
q <- data %>% filter(grepl("^IronSuppSE([1-9])$", Question))

qdata <- merge(q, question_text, by.x = "Question", by.y = "Variable", all.x = TRUE) %>%   
  mutate(
    Statement = sapply(str_split(Question_Text, "[:?]"), function(x) x[2]),
    Statement = ordered(Statement, levels = c(" Restless legs", 
                                              " Dark stool",
                                              " Fatigue",
                                              " Stomach complaints",
                                              " Urge to chew on non-edible products like chalks, ice cubes, pen caps, or raw pasta",
                                              " Headache",
                                              " All of the above",
                                              " None of the above",
                                              " I don't know" )),
    Q = sapply(str_split(Question_Text, "[:?]"), function(x) x[1]),
    answer = ordered(Response, levels=c("TRUE", "FALSE"))
  ) %>% filter(Response == "TRUE") %>% 
  droplevels()


plot_type_2(qdata=qdata, country_colors = country_colors, folder_name=FolderName, VariableName = "Knowledge Iron Supplement Side Effects")
```
